### 2025-01-13 #7658 @editorialbot
[PRE REVIEW]: eitprocessing: a Python package for analysis of Electrical Impedance Tomography data
**Submitting author:** @psomhorst (<a href="http://orcid.org/0000-0003-3490-2080">Peter Somhorst</a>)
**Repository:** https://github.com/EIT-ALIVE/eitprocessing
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.5.2
