### 2025-01-14 3e20386
Add missing DOIs
