+++ paper/paper.bib
@@ -1,10 +1,10 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Peter Somhorst at 2024-09-27 15:57:16 +0200 
+%% Created for Peter Somhorst at 2024-09-27 15:57:16 +0200
 
 
-%% Saved with string encoding Unicode (UTF-8) 
+%% Saved with string encoding Unicode (UTF-8)
 
 
 
@@ -66,7 +66,8 @@
   publisher = {IOP Publishing},
   title     = {Improved filtering methods to suppress cardiovascular contamination in electrical impedance tomography recordings},
   volume    = 45,
-  year      = 2024
+  year      = 2024,
+  doi       = {10.1088/1361-6579/ad46e3}
 }
 
 @article{Bellani2016,
@@ -215,7 +216,8 @@
   keywords = {Electrical impedance tomography; Lung perfusion; Mechanical
               ventilation; Respiratory monitoring; Signal processing;
               Ventilation distribution},
-  language = {en}
+  language = {en},
+  doi      = {10.1186/s13054-024-05173-x}
 }
 
 @article{Wisse2024-sl,
@@ -278,5 +280,6 @@
   keywords = {Advanced respiratory monitoring; Barriers and facilitators;
               Electrical Impedance Tomography (EIT); Esophageal pressure
               measurements (Pes); Implementation; Panel discussion; Survey},
-  language = {en}
+  language = {en},
+  doi      = {10.1186/s40635-024-00686-9}
 }
