### 2025-01-20 917eca8
Clarify determination of file format
