### 2025-01-20 094dd94
Fix type of medibus data to 32 bit float
