### 2025-01-20 098dc69
Move estimation of sample frequency to separate function
