### 2025-01-20 e629943
Bump version: 1.5.2 → 1.6.0
