### 2025-02-11 #7658 @danielskatz
@DaniBodor - I am no longer the track editor. @openjournals/bcm-eics is in charge of the next steps, which should be to find an editor
