+++ paper/paper.bib
@@ -51,7 +51,9 @@
   date-modified = {2024-09-27 15:53:09 +0200},
   publisher     = {New York : McGraw-Hill, Health Professions Division},
   title         = {Principles and practice of intensive care monitoring},
-  year          = {1998}
+  year          = {1998},
+  doi           = {10.1136/thx.53.10.908},
+  url           = {https://doi.org/10.1136/thx.53.10.908}
 }
 
 @article{Wisse2024-wi,
