### 2025-02-21 #7658 @britta-wstnr
Happy to take this on @Kevin-Mattheus-Moerman and @psomhorst! 🌱
I'll go read the paper in more detail now and get back to you within the next few days with how we'll proceed.
I see you already suggested reviewers, thanks for that @psomhorst, and thanks @Kevin-Mattheus-Moerman for prompting the DOI fixes.
