### 2025-03-03 #7658 @britta-wstnr
Hi @drcandacemakedamoore, @tuliofalmeida, and @JanCBrammer! 👋

Would you be willing to review the submission:
**eitprocessing: a Python package for analysis of Electrical Impedance Tomography data**
for [The Journal of Open Source Software (JOSS)](https://joss.theoj.org/)?

The review concerns the software [eitprocessing](https://github.com/EIT-ALIVE/eitprocessing) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.07658/joss.07658/10.21105.joss.07658.pdf) about it. The review happens fully on GitHub.

You can learn more about the the review process and our conflict of interest policy in the reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

If you are available to review this submission, please let me know and I can add you as a reviewer. The review process will start (in a separate GitHub issue) once a sufficient number of reviewers have been found.

Thank you! 🌻
