### 2025-03-03 #7658 @psomhorst
@britta-wstnr As a disclosure, and for your information, @drcandacemakedamoore was in the very early stages part of the team that created this package, and has 33 commits in the repository (https://github.com/openjournals/joss-reviews/issues/7658#issuecomment-2587312501).
