### 2025-03-06 2c9a1f8
Parametrize error message
