### 2025-03-06 00a13a6
Fix potential divide by zero
