### 2025-03-06 07f71dd
Add rudimentary test for rate detection
