### 2025-03-24 #7658 @britta-wstnr
Gentle ping @JanCBrammer, @ajbaird, @ixjlyons - see comment above for a review request from the _Journal of Open Source Software_. Looking forward to hearing from you! 🌱
