### 2025-03-25 db56a90
Add sequence.data object
