### 2025-03-25 18cf99b
Add __contains__ method
