### 2025-03-25 987577a
Add tests for Sequence.data
