### 2025-03-26 e4a968a
Add _collections property
