### 2025-03-26 d735aa8
Add option to add multiple objects at once.
