### 2025-03-26 adbdada
Add magic and accessory methods
