### 2025-03-26 7f3b55c
Add magic and accessory methods
