### 2025-03-29 #7658 @ixjlyons
Apologies for the delayed response. I can't commit to a review in the next month or so, otherwise this looks like something I could review late April / early May if that would be ok.
