### 2025-04-28 #7658 @britta-wstnr
Hi @ixjlyons -- given that I am still searching for reviewers and as per [your comment](https://github.com/openjournals/joss-reviews/issues/7658#issuecomment-2764290794) ... do you think your schedule has cleared up enough to allow for working on this? 🙏
