### 2025-04-28 1d1a58c
Add suggestions for specific interface as alternative when using duplicate labels.
These suggestions only work in Python >=3.11, because they use the Exception.add_note() method.
