### 2025-04-28 17412cf
Replace exception suppression with python version checking
