### 2025-04-28 18646b1
Update ruff version and fix linting issues
