### 2025-04-28 debe0dc
Update ruff version and fix linting issues
