### 2025-04-30 d39a815
Fix pixels being ignored when absolute value negative
Pixels are now skipped when they have no amplitude, i.e. when the standard deviation equals 0.
