### 2025-04-30 07f0c90
Add boolean option to allow pixels with negative amplitude
