### 2025-05-04 367d3e0
Make allow_negative_amplitude class attribute
