### 2025-05-04 98ef67e
Make allow_negative_amplitude True by default
