### 2025-05-04 d889abc
Replace breath_detection_kwargs in TIV and PixelBreath with BreathDetection object
