### 2025-05-04 8382461
Replace breath_detection_kwargs in TIV and PixelBreath with BreathDetection object
