### 2025-05-04 db411d5
Add documentation for new argument
