### 2025-05-04 67da634
Add exception for providing wrong TIV method
