### 2025-05-04 c60bb37
Re-add breath_detection_kwargs argument with DeprecationWarning
