+++ tests/test_parameter_tiv.py
@@ -148,6 +148,17 @@ def test_tiv_initialization():
     assert tiv.breath_detection == BreathDetection()
 
 
+def test_depricated():
+    with pytest.warns(DeprecationWarning):
+        _ = TIV(breath_detection_kwargs={})
+
+    with pytest.raises(TypeError):
+        _ = TIV(breath_detection=BreathDetection(), breath_detection_kwargs={})
+
+    bd_kwargs = {"minimum_duration": 10, "averaging_window_duration": 100.0}
+    assert TIV(breath_detection_kwargs=bd_kwargs).breath_detection == BreathDetection(**bd_kwargs)
+
+
 def test_compute_parameter_type_error():
     """Test that compute_parameter raises TypeError for unsupported data types."""
     tiv = TIV()
