### 2025-05-04 a26e490
Re-add breath_detection_kwargs argument with DeprecationWarning
