+++ tests/test_parameter_tiv.py
@@ -13,7 +13,7 @@ from eitprocessing.datahandling.intervaldata import IntervalData
 from eitprocessing.datahandling.sequence import Sequence
 from eitprocessing.datahandling.sparsedata import SparseData
 from eitprocessing.parameters.tidal_impedance_variation import TIV
-from tests.test_breath_detection import Breath, BreathDetection
+from tests.test_breath_detection import BreathDetection
 
 environment = Path(
     os.environ.get(
