### 2025-05-07 c4d5b68
Replace breath_detection_kwargs in EELI with breath_detection
