### 2025-05-07 febc7a4
Update EELI tests
