### 2025-05-08 adb9e10
Copy potentially non-writeable array
