### 2025-05-09 cc1518a
Replace competing boolean arguments with mode argument
