### 2025-05-09 cee67e3
Update documentation to match new PixelBreath workings
