### 2025-05-09 de4db42
Fix existing test with short data resulting in all None values
