### 2025-05-09 b9609cb
Add test for different phase corrections modes
