+++ tests/test_pixel_breath.py
@@ -380,3 +380,44 @@ def test_with_data(draeger1: Sequence, timpel1: Sequence, pytestconfig: pytest.C
             for breath in filtered_values:
                 # Test whether the indices are in the proper order within a breath
                 assert breath.start_time < breath.middle_time < breath.end_time
+
+
+def test_phase_modes(draeger1: Sequence, pytestconfig: pytest.Config):
+    if pytestconfig.getoption("--cov"):
+        pytest.skip("Skip with option '--cov' so other tests can cover 100%.")
+
+    ssequence = draeger1
+    eit_data = ssequence.eit_data["raw"]
+
+    # reduce the pixel set to middly 'well-behaved' pixels with positive TIV
+    eit_data.pixel_impedance = eit_data.pixel_impedance[:, 10:23, 10:23]
+
+    # flip a single pixel, so the differences between algorithms becomes predictable
+    eit_data.pixel_impedance[:, 6, 6] = -eit_data.pixel_impedance[:, 6, 6]
+
+    cd = ssequence.continuous_data["global_impedance_(raw)"]
+
+    # replace the 'global' data with the sum of the middly pixels
+    cd.values = np.sum(eit_data.pixel_impedance, axis=(1, 2))
+
+    pb_negative_amplitude = PixelBreath(phase_correction_mode="negative amplitude").find_pixel_breaths(eit_data, cd)
+    pb_phase_shift = PixelBreath(phase_correction_mode="phase shift").find_pixel_breaths(eit_data, cd)
+
+    # results are not compared, other than for length; just make sure it runs
+    pb_none = PixelBreath(phase_correction_mode="none").find_pixel_breaths(eit_data, cd)
+
+    assert len(pb_negative_amplitude) == len(pb_phase_shift) == len(pb_none)
+
+    # all breaths, except for the first and last,  should have been detected
+    assert not np.any(np.array(pb_negative_amplitude.values)[1:-1] == None)
+    assert not np.any(np.array(pb_phase_shift.values)[1:-1] == None)
+
+    same_pixel_timing = np.array(pb_negative_amplitude.values) == np.array(pb_phase_shift.values)
+    assert not np.all(same_pixel_timing)
+    assert not np.any(same_pixel_timing[1:-1, 6, 6])  # the single flipped pixel
+    assert np.all(same_pixel_timing[1:-1, :6, :])  # all pixels in the rows above match
+    assert np.all(same_pixel_timing[1:-1, 7:, :])  # all pixels in the rows below match
+    assert np.all(same_pixel_timing[1:-1, :, :6])  # all pixels in the columns to the left match
+    assert np.all(same_pixel_timing[1:-1, :, 7:])  # all pixels in the columns to the right match
+    assert np.all(same_pixel_timing[0, :, :])  # all first values match, because they are all None
+    assert np.all(same_pixel_timing[-1, :, :])  # all last values match, because they are all None
