### 2025-05-12 9f1a08f
Add Mask class
