+++ eitprocessing/roi/__init__.py
@@ -1,0 +1,154 @@
+"""Region of Interest selection and data masking.
+
+This module contains tools for the selection of regions of interest and masking data. The central class of this module
+is `Mask`. Any type of region of interest selection results in a `Mask` object. A mask can be applied to any dataset.
+
+Several default masks have been predefined:
+
+- `VENTRAL_MASK` includes only the first 16 rows;
+- `DORSAL_MASK` includes only the last 16 rows;
+- `RIGHT_MASK` includes only the first 16 columns (NB: right means right lung, which is te left side of the EIT image);
+- `LEFT_MASK` includes only the last 16 columns (NB: left means left lung, which is te right side of the EIT image);
+- `QUADRANT_1_MASK` includes the top right quadrant (NB: right means right lung, which is te left side of the EIT
+  image);
+- `QUADRANT_2_MASK` includes the top left quadrant (NB: left means left lung, which is te right side of the EIT image);
+- `QUADRANT_3_MASK` includes the bottom right quadrant (NB: right means right lung, which is te left side of the EIT
+  image);
+- `QUADRANT_4_MASK` includes the bottom left quadrant (NB: left means left lung, which is te right side of the EIT
+image); - `LAYER_1_MASK` includes only the first 8 rows; - `LAYER_2_MASK` includes only the second set of 8 rows; -
+`LAYER_3_MASK` includes only the third set of 8 rows; - `LAYER_4_MASK` includes only the last 8 rows.
+"""
+
+import dataclasses
+import sys
+from dataclasses import InitVar, dataclass
+from dataclasses import replace as dataclass_replace
+from typing import Literal, Self, TypeVar, overload
+
+import numpy as np
+
+from eitprocessing.datahandling.eitdata import EITData
+from eitprocessing.datahandling.map import PixelMap
+
+T = TypeVar("T", np.ndarray, EITData, PixelMap)
+
+
+@dataclass
+class Mask:
+    """Mask pixels by selecting or weighing them individually.
+
+    A mask is an array with a value for each pixel. Normally, this value is 0 (or False), 1 (or True), and less commonly
+    a value between 0 and 1. When a mask is applied to a dataset, each pixel in that dataset is multiplied by the
+    corresponding masking value. Masking is often used to remove specific pixels by multiplying them by 0. Masking can
+    also be used to weigh pixels, e.g., for a weighted summation. Masking values that are negative or higher than 1 will
+    result in a `ValueError` being raised. You can override this check with `ignore_value_range=True`.
+
+    By default, any masking pixel with the value 0 is converted to `np.nan` ('not a number', NaN). NaN values are
+    different from 0, in that 0 means 'this value is small' and NaN means 'there is no value'. E.g., values outside the
+    region of interest should generally be NaN, not 0. You can prevent the conversion of 0 values to NaN with
+    `zeros_to_nan=False`.
+
+    Masks can be combined by either adding or multiplying them. In either case, the masking values are multiplied to
+    create a new mask.
+
+    Example:
+    ```python
+    assert VENTRAL_MASK + RIGHT_MASK == QUADRANT_1_MASK  # True, quadrant 1 is the ventral part of the right lung
+    assert DORSAL_MASK * LEFT_MASK == QUADRANT_4_MASK  # True, quadrant 4 is the dorsal part of the left lung
+    ```
+
+    """
+
+    mask: np.ndarray
+    description: str | None = None
+    zeros_to_nan: InitVar[bool] = True
+    ignore_value_range: InitVar[bool] = False
+
+    def __post_init__(self, zeros_to_nan: bool, ignore_value_range: bool):
[…]
