### 2025-05-12 9237110
Fix selecting short time with too few breaths
