### 2025-05-12 0c495b4
Fix runtime warning for empty slices
