### 2025-05-12 12be567
Catch or fix deprecation warnings in tests
