### 2025-05-13 51d4e36
Add check for NaN-data before filtering
