### 2025-05-18 e6e1e78
Fix issue where PixelBreath skipped a pixels if a single breath was not detected correctly
