### 2025-05-19 08d8038
Fix bug where breaths were skipped due to lag
