### 2025-05-22 c3d1d88
Detrend data before xcorr
