+++ eitprocessing/features/pixel_breath.py
@@ -198,10 +198,10 @@ class PixelBreath:
                 start_func, middle_func = np.argmin, np.argmax
 
                 cd = np.copy(continuous_data.values)
-                cd -= np.nanmean(cd)
+                cd = signal.detrend(cd, type="linear")
                 pi = np.copy(pixel_impedance[:, row, col])
                 if not np.all(np.isnan(pi)):
-                    pi -= np.nanmean(pixel_impedance[:, row, col])
+                    pi = signal.detrend(pi, type="linear")
 
                 if correct_for_phase_shift:
                     # search for maximum cross correlation within MAX_XCORR_LAG times the average
