### 2025-05-22 7d84206
Set MAX_XCORR_LAG to 1 for peak detection
