### 2025-05-22 161e463
Implement proper peak detection
