### 2025-05-26 37507b2
Add documentation about None values in the output
