### 2025-05-26 0659768
Remove caching from the CI workflow
