### 2025-05-26 2c290f0
Update pip caching
Apparently, we cached the python install location, which is bad
practice. Instead, we should have cached the downloaded packages. This
fixes that.
