+++ .github/actions/install_eitprocessing/action.yml
@@ -41,8 +41,8 @@ runs:
     - uses: actions/cache@v4
       id: cache-python-env
       with:
-        path: ${{ env.pythonLocation }}
-        key: ${{ env.pythonLocation }}-${{ hashFiles('pyproject.toml') }}-[${{ inputs.dependencies }}]
+        path: ~/.cache/pip
+        key: ${{ inputs.python-version }}-${{ hashFiles('pyproject.toml') }}-[${{ inputs.dependencies }}]
     - name: Install eitprocessing with [${{ inputs.dependencies }}] dependencies
       if: steps.cache-python-env.outputs.cache-hit != 'true'
       run: python3 -m pip install ".[${{ inputs.dependencies }}]"
