+++ .github/actions/install_eitprocessing/action.yml
@@ -44,6 +44,5 @@ runs:
         path: ~/.cache/pip
         key: ${{ inputs.python-version }}-${{ hashFiles('pyproject.toml') }}-[${{ inputs.dependencies }}]
     - name: Install eitprocessing with [${{ inputs.dependencies }}] dependencies
-      if: steps.cache-python-env.outputs.cache-hit != 'true'
       run: python3 -m pip install ".[${{ inputs.dependencies }}]"
       shell: bash
