### 2025-05-27 2f3d2c4
Skip warning if all breaths are skipped
