### 2025-05-30 35460e0
Add PixelMap data class
