+++ eitprocessing/datahandling/pixelmap.py
@@ -1,0 +1,240 @@
+from dataclasses import KW_ONLY, dataclass, field
+
+import matplotlib as mpl
+import numpy as np
+from matplotlib import colorbar
+from matplotlib import pyplot as plt
+from matplotlib.axes import Axes
+from matplotlib.colors import CenteredNorm, Colormap, LinearSegmentedColormap, Normalize
+from matplotlib.image import AxesImage
+from matplotlib.ticker import PercentFormatter, ScalarFormatter
+
+ColorType = str | tuple[float, float, float] | tuple[float, float, float, float] | float | Colormap
+
+
+@dataclass(frozen=True)
+class PixelMap:
+    """Map representing a single value for each pixel.
+
+    Common uses are:
+
+    - compliance map, showing the compliance of each pixel;
+    - ODCL map, showing overdistention or collapse values for each pixel;
+    - perfusion map, showing perfusion for each pixel.
+
+    Attributes:
+        values (np.ndarray): 2D array of pixel values.
+        label (str | None): Label for the pixel map.
+        cmap (str | Colormap | None): Colormap for the pixel map.
+        norm (str | Normalize | None): Normalization for the pixel map.
+        facecolor (ColorType): Face color, i.e., the background color for NaN values.
+    """
+
+    values: np.ndarray
+    _: KW_ONLY
+    label: str | None = None
+    cmap: str | Colormap = "viridis"
+    norm: str | Normalize = "linear"
+    facecolor: ColorType = "darkgrey"
+
+    def __post_init__(self):
+        values = np.asarray(self.values, dtype=float)
+        values.flags.writeable = False  # Make the values array immutable
+        object.__setattr__(self, "values", values)
+
+    def imshow(
+        self,
+        colorbar: bool = True,
+        normalize: bool = False,
+        percentage: bool = False,
+        absolute: bool = False,
+        colorbar_kwargs: dict | None = None,
+        facecolor: ColorType | None = None,
+        hide_axes: bool = False,
+        **kwargs,
+    ) -> AxesImage:
+        """Display the pixel map using `imshow`.
+
+        This method is a wrapper around `matplotlib.pyplot.imshow` that provides convenient defaults and formatting
+        options for displaying pixel maps.
+
+        Unless explicitly overridden, the colormap (`cmap`), normalization (`norm`), and axes background color
+        (`facecolor`) are taken from the attributes of the `PixelMap` object.
+
+        If `colorbar` is True, a colorbar is added to the axes. If `normalize` is True, the pixel values are scaled
+        by their maximum value before plotting. The appearance of the colorbar can be modified using `percentage` and
+        `absolute` flags:
+
+        - `percentage=True` displays the colorbar in percentage units (where 1.0 → 100%).
+        - `absolute=True` uses the absolute value of the data for color scaling and labeling.
+
+        If the colormap has underflow or overflow colors (e.g., for negative values), the colorbar will extend
+        accordingly. Additional arguments can be passed to control or override the appearance of the colorbar via
+        `colorbar_kwargs`, which are passed directly to `matplotlib.pyplot.colorbar`.
+
+        If `hide_axes` is True, the axis ticks and labels are hidden (but the axes remain visible to retain background
+        styling such as facecolor).
+
+        Additional keyword arguments are passed directly to `imshow`, allowing for full control over image rendering.
+
+        Args:
+            colorbar (bool): Whether to display a colorbar.
+            normalize (bool): Whether to scale by the maximum value.
+            percentage (bool): Whether to display the colorbar values as a percentage.
[…]
