### 2025-05-30 4a53932
Add PixelMap to the documentation
