### 2025-05-30 60f0373
Fix trying to overwrite non-writeable array
