### 2025-05-30 218a879
Hide axes by default on imshow
