### 2025-05-30 84cb7b4
Move matplotlib formatters to plotting.helpers
