### 2025-05-30 5b06673
Add threshold method to PixelMap
