### 2025-05-31 619d0d4
Add DifferenceMap
