+++ pyproject.toml
@@ -37,7 +37,7 @@ classifiers = [
     "Programming Language :: Python :: 3.13",
 ]
 dependencies = [
-    "matplotlib >= 3.7.1",
+    "matplotlib>=3.10",
     "numpy >= 1.24.2",
     "scipy >= 1.10.1",
     "strenum >= 0.4.10",
