### 2025-05-31 fd1f8e0
Add DifferenceMap
