### 2025-05-31 eaf072a
Import functions from dataclasses instead of the object
