### 2025-05-31 250d0e7
Add convert_to method
