### 2025-06-02 961e5d3
Replace lambda with traditional functions
