### 2025-06-02 6688c62
Move subclass specific factories to class staticmethods
