### 2025-06-02 9a43b09
Fix norm being reused between plots.
