### 2025-06-02 68183c1
Prevent sharing norm between plots if not explicitly set in imshow
