### 2025-06-03 ede3382
Move all plotting parameters to separate class
