+++ eitprocessing/datahandling/pixelmap.py
@@ -1,10 +1,11 @@
 from collections.abc import Callable
 from copy import deepcopy
-from dataclasses import KW_ONLY, asdict, dataclass, field, replace
-from typing import Self, TypeVar
+from dataclasses import KW_ONLY, MISSING, asdict, dataclass, field, replace
+from typing import Self, TypeVar, cast
 
 import matplotlib as mpl
 import numpy as np
+from frozendict import frozendict
 from matplotlib import colorbar
 from matplotlib import pyplot as plt
 from matplotlib.axes import Axes
@@ -29,38 +30,88 @@ def _get_centered_norm() -> CenteredNorm:
 
 
 @dataclass(frozen=True)
+class PlotParameters:
+    """Configuration parameters for plotting pixel maps.
+
+    This class encapsulates visualization settings used when plotting pixel maps, providing a consistent interface for
+    controlling the appearance of plots.
+
+    Attributes:
+        cmap (str | Colormap):
+            The colormap to use for the plot. Can be a string name of a matplotlib colormap or a Colormap instance.
+            Defaults to "viridis".
+        norm (str | Normalize):
+            The normalization to use for the plot. Can be a string or a matplotlib Normalize instance. Defaults to
+            "linear".
+        facecolor (ColorType):
+            The background color for areas with NaN values. Defaults to "darkgrey".
+        colorbar (bool): Whether to display a colorbar. Defaults to True.
+        normalize (bool): Whether to normalize values before plotting. Defaults to False.
+        percentage (bool): Whether to display values as percentages. Defaults to False.
+        absolute (bool): Whether to use absolute values for plotting. Defaults to False.
+        colorbar_kwargs (dict | None): Additional arguments to pass to colorbar creation.
+            Defaults to None.
+        hide_axes (bool): Whether to hide the plot axes. Defaults to True.
+        extra_kwargs (dict): Extra arguments passed to `imshow`. Defaults to an empty dict.
+    """
+
+    cmap: str | Colormap = "viridis"
+    norm: str | Normalize = "linear"
+    facecolor: ColorType = "darkgrey"
+    colorbar: bool = True
+    normalize: bool = False
+    percentage: bool = False
+    absolute: bool = False
+    colorbar_kwargs: frozendict = field(default_factory=frozendict)
+    hide_axes: bool = True
+    extra_kwargs: frozendict = field(default_factory=frozendict)
+
+    def __post_init__(self):
+        for key in ("colorbar_kwargs", "extra_kwargs"):
+            default_factory = self.__dataclass_fields__[key].default_factory
+            default_value = default_factory() if default_factory is not MISSING else None
+            # tell type checker that this is definitely a frozendict
+            default_value = cast("frozendict", default_value)
+
+            merged = default_value | (getattr(self, key) or {})
+
+            object.__setattr__(self, key, merged)
+
+    def replace(self, **changes):
+        if "colorbar_kwargs" in changes:
+            changes["colorbar_kwargs"] = self.colorbar_kwargs | changes["colorbar_kwargs"]
+
+        return replace(self, **changes)
+
+
+@dataclass(frozen=True)
 class PixelMap:
     """Map representing a single value for each pixel.
 
-    Common uses are:
-
-    - compliance map, showing the compliance of each pixel;
-    - ODCL map, showing overdistention or collapse values for each pixel;
-    - perfusion map, showing perfusion for each pixel.
+    For many common cases, specific classes with default plot parameters are available.
 
     Attributes:
         values (np.ndarray): 2D array of pixel values.
         label (str | None): Label for the pixel map.
-        cmap (str | Colormap | None):
[…]
