### 2025-06-03 9a7231c
Add module docs
