### 2025-06-03 420056d
Fix array type hint
