### 2025-06-03 273bba5
Add frozendict to dependencies
