### 2025-06-03 4880a85
Check dimensions of input array
