### 2025-06-03 8c7a7f6
Update PixelMap to have custom __init__ function for proper type checking
We want PixelMap to accept a dict for the plot_parameters argument, but
the final type is PlotParameters. This confuses type checkers if using
the default __init__ method. The same goes for values (np.ndarray
instead of npt.ArrayLike).
