+++ eitprocessing/datahandling/pixelmap.py
@@ -111,7 +111,7 @@ class PlotParameters:
         return replace(self, **changes)
 
 
-@dataclass(frozen=True)
+@dataclass(frozen=True, init=False)
 class PixelMap:
     """Map representing a single value for each pixel.
 
@@ -120,29 +120,38 @@ class PixelMap:
     Attributes:
         values (np.ndarray): 2D array of pixel values.
         label (str | None): Label for the pixel map.
-        plot_parameters (PlotParameters | dict):
+        plot_parameters (PlotParameters):
             Plotting parameters controlling colormap, normalization, colorbar, and other display options. Accepts both a
-            PlotParameters instance or a dict. Subclasses provide their own defaults.
+            PlotParameters instance or a dict, which is converted to PlotParameters during initialization.  Subclasses
+            provide their own defaults.
     """
 
-    values: npt.ArrayLike
+    values: np.ndarray
     _: KW_ONLY
     label: str | None = None
+    plot_parameters: PlotParameters = field(default_factory=PlotParameters)
 
-    plot_parameters: PlotParameters | dict = field(default_factory=PlotParameters)
-
-    def __post_init__(self):
-        values = np.asarray(self.values, dtype=float)
-        if values.ndim != 2:
+    def __init__(
+        self, values: npt.ArrayLike, *, label: str | None = None, plot_parameters: PlotParameters | dict | None = None
+    ):
+        values = np.asarray(values, dtype=float)
+        if values.ndim != 2:  # noqa: PLR2004, ignore hardcoded value
             msg = f"`values` should have 2 dimensions, not {values.ndim}."
             raise ValueError(msg)
 
         values.flags.writeable = False  # Make the values array immutable
         object.__setattr__(self, "values", values)
 
-        if isinstance(self.plot_parameters, dict):
+        object.__setattr__(self, "label", label)
+
+        if plot_parameters is None:
+            plot_parameters = {}
+        if isinstance(plot_parameters, dict):
+            # subclasses define their own version, so grab that if it exists
             plot_parameters_class = getattr(self.__class__, "PlotParameters", PlotParameters)
-            object.__setattr__(self, "plot_parameters", plot_parameters_class(**self.plot_parameters))
+            plot_parameters = plot_parameters_class(**plot_parameters)
+
+        object.__setattr__(self, "plot_parameters", plot_parameters)
 
     def threshold(
         self,
@@ -235,7 +244,7 @@ class PixelMap:
         Returns:
             AxesImage: The image object created by imshow.
         """
-        plot_parameters = cast("PlotParameters", self.plot_parameters)
+        plot_parameters = self.plot_parameters
         normalize = plot_parameters.normalize if normalize is None else normalize
         colorbar = plot_parameters.colorbar if colorbar is None else colorbar
         percentage = plot_parameters.percentage if percentage is None else percentage
@@ -282,7 +291,7 @@ class PixelMap:
     ) -> colorbar.Colorbar:
         """Create a colorbar for the pixel map."""
         colorbar_kwargs = dict(colorbar_kwargs or {})
-        plot_parameters = cast("PlotParameters", self.plot_parameters)
+        plot_parameters = self.plot_parameters
 
         if "format" not in colorbar_kwargs:
             if absolute and percentage:
@@ -348,21 +357,21 @@ class PixelMap:
         Returns:
             Self: A new instance with the replaced attributes.
         """
-        plot_parameters = cast("PlotParameters", self.plot_parameters)
+        plot_parameters = self.plot_parameters
[…]
