### 2025-06-03 d703f0f
Write docs for update method
