### 2025-06-03 3b83ec9
Add tests for PixelMap and subclasses
