+++ tests/test_pixelmap.py
@@ -1,0 +1,266 @@
+from dataclasses import FrozenInstanceError
+
+import frozendict
+import numpy as np
+import pytest
+from matplotlib import pyplot as plt
+from matplotlib.colors import CenteredNorm, Colormap, Normalize
+from matplotlib.ticker import PercentFormatter, ScalarFormatter
+
+from eitprocessing.datahandling.pixelmap import (
+    DifferenceMap,
+    ODCLMap,
+    PerfusionMap,
+    PixelMap,
+    PlotParameters,
+    TIVMap,
+)
+from eitprocessing.plotting.helpers import AbsolutePercentFormatter, AbsoluteScalarFormatter
+
+
+def test_init_values():
+    pm = PixelMap(np.ones((3, 3)))
+    assert isinstance(pm.values, np.ndarray)
+
+    pm = PixelMap([[0]])
+    assert isinstance(pm.values, np.ndarray)
+
+    with pytest.raises(ValueError, match="assignment destination is read-only"):
+        pm.values[0, 0] = 1
+
+    with pytest.raises(ValueError, match="`values` should have 2 dimensions, not 1"):
+        PixelMap([0])
+
+    with pytest.raises(FrozenInstanceError, match="cannot assign to field 'values'"):
+        pm.values = [[1]]
+
+
+def test_init_plot_parameters():
+    pm0 = PixelMap([[0]])
+    pm1 = PixelMap([[0]], plot_parameters={})
+
+    assert pm0.plot_parameters == pm1.plot_parameters
+    assert pm0.plot_parameters is not pm1.plot_parameters
+
+    pm2 = PixelMap([[0]], plot_parameters=PlotParameters(facecolor="white"))
+    pm3 = PixelMap([[0]], plot_parameters={"facecolor": "white"})
+
+    assert pm2.plot_parameters != pm0.plot_parameters
+    assert pm2.plot_parameters == pm3.plot_parameters
+
+    pm4 = pm0.update(plot_parameters=PlotParameters(facecolor="white"))
+    pm5 = pm0.update(plot_parameters={"facecolor": "white"})
+
+    assert pm4.plot_parameters != pm0.plot_parameters
+    assert pm4.plot_parameters == pm2.plot_parameters
+    assert pm4.plot_parameters == pm5.plot_parameters
+
+    pm6 = pm2.update(plot_parameters={"cmap": "inferno"})
+    pm7 = pm2.update(plot_parameters=PlotParameters(cmap="inferno"))
+    pm8 = pm2.update(plot_parameters=pm2.plot_parameters.update(cmap="inferno"))
+
+    # when passing a dict as plot_parameters, plot_parameters is *updated*, not *replaced*
+    # pm6 works the same as pm8 internally
+    # when passing a PlotParameters object, the entire object is replaced
+    # pm7 should probably almost never be used
+    assert pm6.plot_parameters != pm7.plot_parameters
+    assert pm7.plot_parameters.facecolor != "white"
+    assert pm6.plot_parameters == pm8.plot_parameters
+
+    with pytest.raises(TypeError, match=r"PlotParameters.__init__\(\) got an unexpected keyword argument"):
+        _ = PixelMap([[0]], plot_parameters={"non_existing": None})
+
+    # normally, colorbar kwargs is empty
+    assert pm6.plot_parameters.colorbar_kwargs == {}
+    pm9 = PixelMap([[0]], plot_parameters={"colorbar_kwargs": {"key": "value"}})
[…]
