### 2025-06-04 b1ae6c4
Add magic methods for adding, subtracting, multiplying and dividing PixelMaps
