### 2025-06-04 8af44b7
Add tests for PixelMap math
