### 2025-06-04 c9b5cde
Add documentation and restructure API docs
The API Documentation page was becoming a very long page. I restructured
it to have a single page per module.
