+++ docs/api/datacontainers.md
@@ -1,25 +1,0 @@
-# Data containers
-
-::: eitprocessing.datahandling.sequence.Sequence
-
-::: eitprocessing.datahandling.eitdata.EITData
-
-::: eitprocessing.datahandling.continuousdata.ContinuousData
-
-::: eitprocessing.datahandling.sparsedata.SparseData
-
-::: eitprocessing.datahandling.intervaldata.IntervalData
-
-::: eitprocessing.datahandling.pixelmap.PixelMap
-
-::: eitprocessing.datahandling.pixelmap.TIVMap
-
-::: eitprocessing.datahandling.pixelmap.ODCLMap
-
-::: eitprocessing.datahandling.pixelmap.DifferenceMap
-
-::: eitprocessing.datahandling.datacollection.DataCollection
-
-::: eitprocessing.datahandling.breath.Breath
-
-::: eitprocessing.datahandling.event.Event
