+++ docs/api/datacontainers/datacollection.md
@@ -1,0 +1,1 @@
+::: eitprocessing.datahandling.datacollection
