+++ docs/api/datacontainers/sequence.md
@@ -1,0 +1,1 @@
+::: eitprocessing.datahandling.sequence
