+++ docs/api/datacontainers/event.md
@@ -1,0 +1,1 @@
+::: eitprocessing.datahandling.event
