+++ docs/api/datacontainers/pixelmap.md
@@ -1,0 +1,11 @@
+::: eitprocessing.datahandling.pixelmap
+    options:
+        members:
+        - PixelMap
+        - DifferenceMap
+        - TIVMap
+        - ODCLMap
+        - PerfusionMap
+        - PendelluftMap
+        - SignedPendelluftMap
+        - PlotParameters
