### 2025-06-04 2fd15fe
Make PlotParameters init keyword only
