### 2025-06-04 34f5f5d
Update Self from typing_extensions, for Python 3.10
