### 2025-06-05 08a18bc
Remove impossible condition check
If __rsub__ is called, other is by definition not a PixelMap.
