### 2025-06-05 60ebdba
Add from_mean method and tests
