### 2025-06-05 db5d488
Add from_mean method and tests
