### 2025-06-18 c8eaafa
Rename update to __replace__ to support copy.replace()
copy.replace() is a new feature in Python 3.13, that introduces a common way to update existing objects, similar to dataclasses.replace. __replace__ implements this for PixelMap and PlotParameters.
