### 2025-07-10 02fb1c5
Move PixelMap plotting logic to separate class
