+++ tests/test_pixelmap.py
@@ -15,7 +15,6 @@ from eitprocessing.datahandling.pixelmap import (
     PendelluftMap,
     PerfusionMap,
     PixelMap,
-    PlotParameters,
     SignedPendelluftMap,
     TIVMap,
 )
@@ -52,7 +51,8 @@ def test_update():
     assert pm2.label == "test"
     assert pm2 == pm0.update(values=[[1]], label="test")
 
-    pp0 = PlotParameters()
+    pp0 = PixelMap.PlotParameters()
+    assert isinstance(pp0, PixelMap.PlotParameters)
     pp1 = copy.replace(pp0, cmap="foo")
     assert pp1.cmap == "foo"
     assert pp1 == pp0.update(cmap="foo")
@@ -62,49 +62,50 @@ def test_init_plot_parameters():
     pm0 = PixelMap([[0]])
     pm1 = PixelMap([[0]], plot_parameters={})
 
-    assert pm0.plot_parameters == pm1.plot_parameters
-    assert pm0.plot_parameters is not pm1.plot_parameters
+    assert pm0._plot_parameters == pm1._plot_parameters
+    assert pm0._plot_parameters is not pm1._plot_parameters
 
-    pm2 = PixelMap([[0]], plot_parameters=PlotParameters(facecolor="white"))
+    pm2 = PixelMap([[0]], plot_parameters=PixelMap.PlotParameters(facecolor="white"))
     pm3 = PixelMap([[0]], plot_parameters={"facecolor": "white"})
 
-    assert pm2.plot_parameters != pm0.plot_parameters
-    assert pm2.plot_parameters == pm3.plot_parameters
+    assert pm2._plot_parameters != pm0._plot_parameters
+    assert pm2._plot_parameters == pm3._plot_parameters
 
-    pm4 = pm0.update(plot_parameters=PlotParameters(facecolor="white"))
+    pm4 = pm0.update(plot_parameters=PixelMap.PlotParameters(facecolor="white"))
     pm5 = pm0.update(plot_parameters={"facecolor": "white"})
 
-    assert pm4.plot_parameters != pm0.plot_parameters
-    assert pm4.plot_parameters == pm2.plot_parameters
-    assert pm4.plot_parameters == pm5.plot_parameters
+    assert pm4._plot_parameters != pm0._plot_parameters
+    assert pm4._plot_parameters == pm2._plot_parameters
+    assert pm4._plot_parameters == pm5._plot_parameters
 
     pm6 = pm2.update(plot_parameters={"cmap": "inferno"})
-    pm7 = pm2.update(plot_parameters=PlotParameters(cmap="inferno"))
-    pm8 = pm2.update(plot_parameters=pm2.plot_parameters.update(cmap="inferno"))
+    pm7 = pm2.update(plot_parameters=PixelMap.PlotParameters(cmap="inferno"))
+    pm8 = pm2.update(plot_parameters=pm2._plot_parameters.update(cmap="inferno"))
 
     # when passing a dict as plot_parameters, plot_parameters is *updated*, not *replaced*
     # pm6 works the same as pm8 internally
     # when passing a PlotParameters object, the entire object is replaced
     # pm7 should probably almost never be used
-    assert pm6.plot_parameters != pm7.plot_parameters
-    assert pm7.plot_parameters.facecolor != "white"
-    assert pm6.plot_parameters == pm8.plot_parameters
+    assert pm6._plot_parameters != pm7._plot_parameters
+    assert pm7._plot_parameters.facecolor != "white"
[…]
