+++ eitprocessing/datahandling/pixelmap.py
@@ -28,108 +28,28 @@ plotting parameter).
 from __future__ import annotations
 
 import warnings
-from copy import deepcopy
-from dataclasses import KW_ONLY, MISSING, asdict, dataclass, field, replace
-from typing import TYPE_CHECKING, TypeVar, cast
+from dataclasses import KW_ONLY, InitVar, asdict, dataclass, field, replace
+from typing import TYPE_CHECKING, TypeVar
 
 import matplotlib as mpl
 import numpy as np
 from frozendict import frozendict
-from matplotlib import colorbar
-from matplotlib import pyplot as plt
 from matplotlib.colors import CenteredNorm, Colormap, LinearSegmentedColormap, Normalize
-from matplotlib.ticker import PercentFormatter
 from numpy import typing as npt
 from typing_extensions import Self
 
-from eitprocessing.plotting.helpers import AbsolutePercentFormatter, AbsoluteScalarFormatter
+from eitprocessing.plotting.pixelmap_plotting import (
+    ColorType,
+    Parameters,
+    PixelMapPlotting,
+    _get_centered_norm,
+    _get_zero_norm,
+)
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Sequence
 
-    from matplotlib.axes import Axes
-    from matplotlib.image import AxesImage
-
-ColorType = str | tuple[float, float, float] | tuple[float, float, float, float] | float | Colormap
-
 T = TypeVar("T", bound="PixelMap")
-
-
-def _get_zero_norm() -> Normalize:
-    return Normalize(vmin=0)
-
-
-def _get_centered_norm() -> CenteredNorm:
-    return CenteredNorm(vcenter=0)
-
-
-@dataclass(frozen=True, kw_only=True)
-class PlotParameters:
-    """Configuration parameters for plotting pixel maps.
-
-    This class encapsulates visualization settings used when plotting pixel maps, providing a consistent interface for
-    controlling the appearance of plots.
-
-    Attributes:
-        cmap (str | Colormap):
-            The colormap to use for the plot. Can be a string name of a matplotlib colormap or a Colormap instance.
-            Defaults to "viridis".
-        norm (str | Normalize):
-            The normalization to use for the plot. Can be a string or a matplotlib Normalize instance. Defaults to
-            "linear".
-        facecolor (ColorType):
-            The background color for areas with NaN values. Defaults to "darkgrey".
-        colorbar (bool): Whether to display a colorbar. Defaults to True.
-        normalize (bool): Whether to normalize values before plotting. Defaults to False.
-        percentage (bool): Whether to display values as percentages. Defaults to False.
-        absolute (bool): Whether to use absolute values for plotting. Defaults to False.
-        colorbar_kwargs (dict | None): Additional arguments to pass to colorbar creation.
-            Defaults to None.
-        hide_axes (bool): Whether to hide the plot axes. Defaults to True.
-        extra_kwargs (dict): Extra arguments passed to `imshow`. Defaults to an empty dict.
-    """
-
-    cmap: str | Colormap = "viridis"
-    norm: str | Normalize = "linear"
-    facecolor: ColorType = "darkgrey"
-    colorbar: bool = True
-    normalize: bool = False
-    percentage: bool = False
-    absolute: bool = False
-    colorbar_kwargs: frozendict = field(default_factory=frozendict)
-    hide_axes: bool = True
-    extra_kwargs: frozendict = field(default_factory=frozendict)
-
-    def __post_init__(self):
-        for key in ("colorbar_kwargs", "extra_kwargs"):
-            default_factory = self.__dataclass_fields__[key].default_factory
-            default_value = default_factory() if default_factory is not MISSING else None
-
-            # tell type checker that this is definitely a frozendict
-            default_value = cast("frozendict", default_value)
-
[…]
