### 2025-07-14 6431491
Move all PlotParameters to plotting module
