+++ eitprocessing/datahandling/pixelmap.py
@@ -31,23 +31,14 @@ import warnings
 from dataclasses import KW_ONLY, InitVar, asdict, dataclass, field, replace
 from typing import TYPE_CHECKING, TypeVar
 
-import matplotlib as mpl
 import numpy as np
-from frozendict import frozendict
-from matplotlib.colors import CenteredNorm, Colormap, LinearSegmentedColormap, Normalize
 from numpy import typing as npt
 from typing_extensions import Self
 
-from eitprocessing.plotting.pixelmap import (
-    ColorType,
-    Parameters,
-    PixelMapPlotting,
-    _get_centered_norm,
-    _get_zero_norm,
-)
-
 if TYPE_CHECKING:
     from collections.abc import Callable, Sequence
+
+    from eitprocessing.plotting.pixelmap import PixelMapPlotParameters, PixelMapPlotting
 
 T = TypeVar("T", bound="PixelMap")
 
@@ -72,51 +63,18 @@ class PixelMap:
             image.
     """
 
-    @dataclass(frozen=True, kw_only=True)
-    class PlotParameters(Parameters):
-        """Configuration parameters for plotting pixel maps.
-
-        This class encapsulates visualization settings used when plotting pixel maps, providing a consistent interface
-        for controlling the appearance of plots.
-
-        Attributes:
-            cmap (str | Colormap):
-                The colormap to use for the plot. Can be a string name of a matplotlib colormap or a Colormap instance.
-                Defaults to "viridis".
-            norm (str | Normalize):
-                The normalization to use for the plot. Can be a string or a matplotlib Normalize instance. Defaults to
-                "linear".
-            facecolor (ColorType):
-                The background color for areas with NaN values. Defaults to "darkgrey".
-            colorbar (bool): Whether to display a colorbar. Defaults to True.
-            normalize (bool): Whether to normalize values before plotting. Defaults to False.
-            percentage (bool): Whether to display values as percentages. Defaults to False.
-            absolute (bool): Whether to use absolute values for plotting. Defaults to False.
-            colorbar_kwargs (dict | None): Additional arguments to pass to colorbar creation.
-                Defaults to None.
-            hide_axes (bool): Whether to hide the plot axes. Defaults to True.
-            extra_kwargs (dict): Extra arguments passed to `imshow`. Defaults to an empty dict.
-        """
-
-        cmap: str | Colormap = "viridis"
-        norm: str | Normalize = "linear"
-        facecolor: ColorType = "darkgrey"
-        colorbar: bool = True
-        normalize: bool = False
-        percentage: bool = False
-        absolute: bool = False
-        colorbar_kwargs: frozendict = field(default_factory=frozendict)
-        hide_axes: bool = True
-        extra_kwargs: frozendict = field(default_factory=frozendict)
-
     values: np.ndarray
     _: KW_ONLY
     label: str | None = None
-    plot_parameters: InitVar[PlotParameters]
-    _plot_parameters: PlotParameters = field(init=False, repr=False)
+    plot_parameters: InitVar[PixelMapPlotParameters | dict | None]
+    _plot_parameters: PixelMapPlotParameters = field(init=False, repr=False)
 
     def __init__(
-        self, values: npt.ArrayLike, *, label: str | None = None, plot_parameters: PlotParameters | dict | None = None
+        self,
+        values: npt.ArrayLike,
+        *,
+        label: str | None = None,
+        plot_parameters: PixelMapPlotParameters | dict | None = None,
     ):
         values = np.asarray(values, dtype=float)
         if values.ndim != 2:  # noqa: PLR2004, ignore hardcoded value
[…]
