### 2025-07-14 285843f
Move tests to more logical functions
