### 2025-07-14 56a11e3
Make plotparameters an instance
