### 2025-07-14 f5f27bf
Add tests for changing plot parameter defaults
