### 2025-07-14 9b02f9a
Add set and reset functions for pixelmap plot parameter registry
