+++ eitprocessing/plotting/pixelmap.py
@@ -8,20 +8,20 @@ The plotting parameters are defined in the `PixelMapPlotParameters` class and it
 types. PIXELMAP_PLOT_PARAMETERS_REGISTRY is a registry that maps pixel map types to their respective plotting
 parameters. At initialization of the `PixelMap` class, the appropriate plotting parameters are copied from the registry.
 The `get_pixelmap_plot_parameters` function retrieves the plotting parameters for a specific pixel map instance or type.
-The registry can be updated using the `set_pixelmap_plot_parameters` function, which allows for changing the defaults
+The registry can be updated using the `set_pixelmap_plot_config` function, which allows for changing the defaults
 for specific pixel map types or all types at once. The registry can be reset to hardcoded defaults using the
 `reset_pixelmap_plot_parameters`.
 
 Examples:
     >>> from eitprocessing.datahandling.pixelmap import PixelMap, TIVMap
-    >>> from eitprocessing.plotting.pixelmap import get_pixelmap_plot_parameters, set_pixelmap_plot_parameters
+    >>> from eitprocessing.plotting.pixelmap import get_pixelmap_plot_parameters, set_pixelmap_plot_config
 
     # Get default parameters for a pixel map instance
     >>> pixel_map = PixelMap(values=np.random.rand(10, 10))
     >>> params = get_pixelmap_plot_parameters(pixel_map)
 
     # Update parameters for TIVMap
-    >>> set_pixelmap_plot_parameters(TIVMap, cmap="plasma")
+    >>> set_pixelmap_plot_config(TIVMap, cmap="plasma")
 
     # Reset all parameters to hardcoded defaults
     >>> reset_pixelmap_plot_parameters()
@@ -65,8 +65,8 @@ def _get_field_type(field_: Field[T], cls: type) -> type[T]:
 
 
 @dataclass(frozen=True, kw_only=True)
-class Parameters:
-    """Base class for parameters."""
+class Config:
+    """Base class for configuration."""
 
     def __post_init__(self):
         for field_ in fields(self):
@@ -78,7 +78,7 @@ class Parameters:
                 object.__setattr__(self, field_.name, frozendict(merged))
 
     def __replace__(self, /, **changes) -> Self:
-        """Return a copy of the of the PlotParameters instance replacing attributes.
+        """Return a copy of the of the Config instance replacing attributes.
 
         Similar to dataclass.replace(), but with special handling of `colorbar_kwargs`. `colorbar_kwargs` is updated
         with the provided dictionary, rather than replaced.
@@ -114,9 +114,9 @@ class PixelMapPlotting:
     pixel_map: PixelMap = field(compare=False, repr=False)
 
     @property
-    def parameters(self) -> "PixelMapPlotParameters":
-        """Plotting parameters for the pixel map."""
-        return self.pixel_map._plot_parameters  # noqa: SLF001
+    def config(self) -> "PixelMapPlotConfig":
+        """Plotting configuration for the pixel map."""
+        return self.pixel_map._plot_config  # noqa: SLF001
 
     def imshow(
         self,
@@ -169,20 +169,20 @@ class PixelMapPlotting:
         Returns:
             AxesImage: The image object created by imshow.
         """
-        normalize = self.parameters.normalize if normalize is None else normalize
-        colorbar = self.parameters.colorbar if colorbar is None else colorbar
[…]
