+++ tests/test_pixelmap.py
@@ -20,11 +20,11 @@ from eitprocessing.datahandling.pixelmap import (
 )
 from eitprocessing.plotting.helpers import AbsolutePercentFormatter, AbsoluteScalarFormatter
 from eitprocessing.plotting.pixelmap import (
-    PIXELMAP_PLOT_PARAMETERS_REGISTRY,
-    PixelMapPlotParameters,
-    TIVMapPlotParameters,
-    reset_pixelmap_plot_parameters,
-    set_pixelmap_plot_parameters,
+    PIXELMAP_PLOT_CONFIG_REGISTRY,
+    PixelMapPlotConfig,
+    TIVMapPlotConfig,
+    reset_pixelmap_plot_config,
+    set_pixelmap_plot_config,
 )
 
 
@@ -63,69 +63,69 @@ def test_init_plot_parameters():
     pm0 = PixelMap([[0]])
     pm1 = PixelMap([[0]], plot_parameters={})
 
-    assert pm0._plot_parameters is pm0.plotting.parameters
-    assert pm1._plot_parameters is pm1.plotting.parameters
-    assert isinstance(pm0.plotting.parameters, PixelMapPlotParameters)
-    assert isinstance(pm1.plotting.parameters, PixelMapPlotParameters)
+    assert pm0._plot_config is pm0.plotting.config
+    assert pm1._plot_config is pm1.plotting.config
+    assert isinstance(pm0.plotting.config, PixelMapPlotConfig)
+    assert isinstance(pm1.plotting.config, PixelMapPlotConfig)
 
-    assert pm0.plotting.parameters == pm1.plotting.parameters
-    assert pm0.plotting.parameters is not pm1.plotting.parameters
+    assert pm0.plotting.config == pm1.plotting.config
+    assert pm0.plotting.config is not pm1.plotting.config
 
-    pm2 = PixelMap([[0]], plot_parameters=PixelMapPlotParameters(facecolor="white"))
+    pm2 = PixelMap([[0]], plot_parameters=PixelMapPlotConfig(facecolor="white"))
     pm3 = PixelMap([[0]], plot_parameters={"facecolor": "white"})
 
-    assert isinstance(pm2.plotting.parameters, PixelMapPlotParameters)
+    assert isinstance(pm2.plotting.config, PixelMapPlotConfig)
 
-    assert pm2.plotting.parameters != pm0.plotting.parameters
-    assert pm2.plotting.parameters == pm3.plotting.parameters
+    assert pm2.plotting.config != pm0.plotting.config
+    assert pm2.plotting.config == pm3.plotting.config
 
     with pytest.raises(TypeError, match=r"PlotParameters.__init__\(\) got an unexpected keyword argument"):
         _ = PixelMap([[0]], plot_parameters={"non_existing": None})
 
     # normally, colorbar kwargs is empty
-    assert pm0.plotting.parameters.colorbar_kwargs == {}
+    assert pm0.plotting.config.colorbar_kwargs == {}
     pm3 = PixelMap([[0]], plot_parameters={"colorbar_kwargs": {"key": "value"}})
-    assert pm3.plotting.parameters.colorbar_kwargs == {"key": "value"}
-    assert isinstance(pm3.plotting.parameters.colorbar_kwargs, frozendict.frozendict)
+    assert pm3.plotting.config.colorbar_kwargs == {"key": "value"}
+    assert isinstance(pm3.plotting.config.colorbar_kwargs, frozendict.frozendict)
 
 
 def test_update_plot_parameters():
-    pp0 = PixelMapPlotParameters()
[…]

[…]
