### 2025-07-18 eefdab5
Add comment about mutability to PIxelMap class.
