### 2025-07-18 37131eb
Update eitprocessing/datahandling/pixelmap.py
Co-authored-by: JulietteFrancovich <78535051+JulietteFrancovich@users.noreply.github.com>
