### 2025-07-22 3caeb95
When using __replace__ set label to None by default
