### 2025-07-22 2d63023
Add normalize() method, replacing imshow argument and plotting parameter
