+++ eitprocessing/datahandling/pixelmap.py
@@ -30,7 +30,7 @@ from __future__ import annotations
 import warnings
 from copy import deepcopy
 from dataclasses import KW_ONLY, MISSING, asdict, dataclass, field, replace
-from typing import TYPE_CHECKING, TypeVar, cast
+from typing import TYPE_CHECKING, Literal, TypeVar, cast
 
 import matplotlib as mpl
 import numpy as np
@@ -80,7 +80,6 @@ class PlotParameters:
         facecolor (ColorType):
             The background color for areas with NaN values. Defaults to "darkgrey".
         colorbar (bool): Whether to display a colorbar. Defaults to True.
-        normalize (bool): Whether to normalize values before plotting. Defaults to False.
         percentage (bool): Whether to display values as percentages. Defaults to False.
         absolute (bool): Whether to use absolute values for plotting. Defaults to False.
         colorbar_kwargs (dict | None): Additional arguments to pass to colorbar creation.
@@ -93,7 +92,6 @@ class PlotParameters:
     norm: str | Normalize = "linear"
     facecolor: ColorType = "darkgrey"
     colorbar: bool = True
-    normalize: bool = False
     percentage: bool = False
     absolute: bool = False
     colorbar_kwargs: frozendict = field(default_factory=frozendict)
@@ -178,6 +176,85 @@ class PixelMap:
 
         object.__setattr__(self, "plot_parameters", plot_parameters)
 
+    def normalize(
+        self,
+        mode: Literal["zero-based", "symmetric", "maximum", "reference"] = "zero-based",
+        reference: float | None = None,
+        **kwargs,
+    ) -> Self:
+        """Normalize the pixel map values.
+
+        Creates a copy of the pixel map with normalized values. Four normalization modes are available.
+
+        The default mode is "zero-based", which normalizes the values to the range [0, 1] by subtracting the minimum
+        value and dividing by the new maximum value. This ensures the lowest resulting value is 0 and the highest
+        resulting value is 1.
+
+        The mode "symmetric" divides the values by the maximum absolute value, resulting in a range of [-1, 1].
+
+        The mode "maximum" divides the values by the maximum value. If all values are positive, this normalizes
+        them to the range [0, 1] without shifting the minimum value to zero. Negative values are scaled by the maximum
+        values, and might result in values lower than -1.
+
+        The mode "reference" is similar to "maximum", except it divides by a user-defined reference value. A reference
+        value must be provided. Resulting values can fall outside the range [-1, 1].
+
+        Examples:
+        ```
+        >>> PixelMap([[1, 3, 5]]).normalize()  # Default is zero-based normalization
+        PixelMap(values=array([[0. , 0.5, 1. ]]), ...)
+
+        >>> PixelMap([[1, 3, 5]]).normalize(mode="maximum")
+        PixelMap(values=array([[0.2, 0.6, 1. ]]), ...)
+
+        >>> PixelMap([[-8, -1, 2]]).normalize()
+        PixelMap(values=array([[0. , 0.7, 1. ]]), ...)
+
+        >>> PixelMap([[-8, -1, 2]]).normalize(mode="symmetric")
+        PixelMap(values=array([[-1.   , -0.125,  0.25 ]]), ...)
+
+        >>> PixelMap([[-8, -1, 2]]).normalize(mode="reference", reference=4)
+        PixelMap(values=array([[-2.  , -0.25,  0.5 ]]), ...)
+        ```
+
+        Args:
+            mode (Literal["zero-based", "symmetric", "maximum", "reference"]):
+                The normalization mode to use. Defaults to "zero-based".
+            reference (float | None):
+                The reference value to use for normalization in "reference" mode.
+            kwargs (dict):
+                Additional keyword arguments to pass to the new PixelMap instance.
+        """
+        if reference is not None and mode != "reference":
[…]
