### 2025-07-24 d016388
Improve normalize method
