### 2025-07-24 fa87653
Add tests for PixelMap.normalize
