### 2025-07-24 d34446b
Add warnings for all negative or all nan values
