### 2025-07-25 ba8ed5c
Add tests for warning suppression
