### 2025-07-25 70ddd5a
Add stage property
