### 2025-07-25 c9d20f8
Add temporary PixelMap dummy class
