+++ eitprocessing/roi/__init__.py
@@ -1,29 +1,30 @@
-"""Region of Interest selection and data masking.
+"""Region of Interest selection and pixel masking.
 
-This module contains tools for the selection of regions of interest and masking data. The central class of this module
-is `Mask`. Any type of region of interest selection results in a `Mask` object. A mask can be applied to any dataset.
+This module contains tools for the selection of regions of interest and masking pixel data. The central class of this
+module is `PixelMask`. Any type of region of interest selection results in a `PixelMask` object. A mask can be applied
+to any pixel dataset (EITData, PixelMap) with the same shape.
 
-Several default masks have been predefined:
+Several default masks have been predefined. NB: the right lung is to the left side of the EIT image and vice versa.
 
 - `VENTRAL_MASK` includes only the first 16 rows;
 - `DORSAL_MASK` includes only the last 16 rows;
-- `RIGHT_MASK` includes only the first 16 columns (NB: right means right lung, which is te left side of the EIT image);
-- `LEFT_MASK` includes only the last 16 columns (NB: left means left lung, which is te right side of the EIT image);
-- `QUADRANT_1_MASK` includes the top right quadrant (NB: right means right lung, which is te left side of the EIT
-  image);
-- `QUADRANT_2_MASK` includes the top left quadrant (NB: left means left lung, which is te right side of the EIT image);
-- `QUADRANT_3_MASK` includes the bottom right quadrant (NB: right means right lung, which is te left side of the EIT
-  image);
-- `QUADRANT_4_MASK` includes the bottom left quadrant (NB: left means left lung, which is te right side of the EIT
-image); - `LAYER_1_MASK` includes only the first 8 rows; - `LAYER_2_MASK` includes only the second set of 8 rows; -
-`LAYER_3_MASK` includes only the third set of 8 rows; - `LAYER_4_MASK` includes only the last 8 rows.
+- `RIGHT_LUNG_MASK` includes only the first 16 columns;
+- `LEFT_LUNG_MASK` includes only the last 16 columns;
+- `QUADRANT_1_MASK` includes the top right quadrant;
+- `QUADRANT_2_MASK` includes the top left quadrant;
+- `QUADRANT_3_MASK` includes the bottom right quadrant;
+- `QUADRANT_4_MASK` includes the bottom left quadrant;
+- `LAYER_1_MASK` includes only the first 8 rows;
+- `LAYER_2_MASK` includes only the second set of 8 rows;
+- `LAYER_3_MASK` includes only the third set of 8 rows;
+- `LAYER_4_MASK` includes only the last 8 rows.
 """
 
 import dataclasses
 import sys
-from dataclasses import InitVar, dataclass
+from dataclasses import InitVar, dataclass, field
 from dataclasses import replace as dataclass_replace
-from typing import Literal, Self, TypeVar, overload
+from typing import Self, TypeVar, overload
 
 import numpy as np
 
@@ -39,67 +40,82 @@ class PixelMap:  # noqa: D101
 T = TypeVar("T", np.ndarray, EITData, PixelMap)
 
 
-@dataclass
-class Mask:
+@dataclass(frozen=True)
+class PixelMask:
     """Mask pixels by selecting or weighing them individually.
 
-    A mask is an array with a value for each pixel. Normally, this value is 0 (or False), 1 (or True), and less commonly
-    a value between 0 and 1. When a mask is applied to a dataset, each pixel in that dataset is multiplied by the
-    corresponding masking value. Masking is often used to remove specific pixels by multiplying them by 0. Masking can
-    also be used to weigh pixels, e.g., for a weighted summation. Masking values that are negative or higher than 1 will
[…]
