### 2025-07-25 eaa2251
Add tests for PixelMask
