+++ tests/test_pixelmask.py
@@ -1,0 +1,281 @@
+import numpy as np
+import pytest
+
+from eitprocessing.datahandling.sequence import Sequence
+from eitprocessing.roi import (
+    DORSAL_MASK,
+    LAYER_1_MASK,
+    LAYER_2_MASK,
+    LAYER_3_MASK,
+    LAYER_4_MASK,
+    LEFT_LUNG_MASK,
+    QUADRANT_1_MASK,
+    QUADRANT_2_MASK,
+    QUADRANT_3_MASK,
+    QUADRANT_4_MASK,
+    RIGHT_LUNG_MASK,
+    VENTRAL_MASK,
+    PixelMask,
+)
+
+
+def test_pixelmask_init_basic(): ...
+
+
+def test_pixelmask_init_with_boolean_array():
+    values = np.random.default_rng().random((10, 10)) > 0.5
+    assert np.all((values == True) | (values == False))  # noqa: E712
+
+    mask = PixelMask(values)
+
+    assert np.all((np.isnan(mask.mask)) | (mask.mask == 1.0))
+    assert np.array_equal(np.isnan(mask.mask), ~values)
+    assert np.array_equal(mask.mask == 1.0, values)
+
+
+def test_pixelmask_init_with_integer_array():
+    values = np.random.default_rng().integers(0, 2, (10, 10))
+    mask = PixelMask(values)
+
+    assert np.all((np.isnan(mask.mask)) | (mask.mask == 1.0))
+    assert np.array_equal(np.isnan(mask.mask), values == 0)
+    assert np.array_equal(mask.mask == 1.0, values == 1)
+
+
+def test_pixelmask_init_with_float_array_non_weighted():
+    values = np.random.default_rng().random((10, 10)).round()
+    assert np.all((values == 0.0) | (values == 1.0))
+
+    mask = PixelMask(values)
+
+    assert np.all((np.isnan(mask.mask)) | (mask.mask == 1.0))
+    assert np.array_equal(np.isnan(mask.mask), values == 0.0)
+    assert np.array_equal(mask.mask == 1.0, values == 1.0)
+
+
+def test_pixelmask_init_with_float_array_weighted():
+    values = np.random.default_rng().random((10, 10))
+    values[values < 0.5] = 0.0
+
+    mask = PixelMask(values)
+
+    assert np.array_equal(np.isnan(mask.mask), values < 0.5)  # NaN for values < 0.5
+    assert np.array_equal(~np.isnan(mask.mask), values >= 0.5)  # non-NaN for values >= 0.5
+    assert np.array_equal(mask.mask[~np.isnan(mask.mask)], values[~np.isnan(mask.mask)])  # non-NaN values are unaltered
+
+
+def test_pixelmask_init_with_list():
+    values = [[1, 0, 1], [0, 1, 0]]
+    mask = PixelMask(values)
+
+    assert np.array_equal(mask.mask, np.array([[1.0, np.nan, 1.0], [np.nan, 1.0, np.nan]]), equal_nan=True)
+
+
+def test_pixelmask_init_keep_zeros_true():
+    values = np.random.default_rng().random((10, 10))
+    values[values < 0.5] = 0.0
+
+    mask = PixelMask(values, keep_zeros=True)
+
+    assert np.array_equal(mask.mask == 0.0, values < 0.5)  # NaN for values < 0.5
+    assert np.array_equal(mask.mask[mask.mask >= 0.5], values[mask.mask >= 0.5])  # non-NaN values are unaltered
+
+
+def test_pixelmask_init_invalid_dtype():
[…]
