### 2025-07-25 7426c1f
Revert "Add temporary PixelMap dummy class"
This reverts commit c9d20f855c3e16c97adf04d78e40b2c453a9b0c3.
