### 2025-07-25 b1c526c
Replace PixelMap.threshold function with create_mask_from_treshold()
