### 2025-07-31 b3ba5d2
Rename allows_negative_values to allow_negative_values
