### 2025-07-31 5e3d34f
Move plotting logic to separate module
