+++ tests/test_pixelmap.py
@@ -16,11 +16,16 @@ from eitprocessing.datahandling.pixelmap import (
     PendelluftMap,
     PerfusionMap,
     PixelMap,
-    PlotParameters,
     SignedPendelluftMap,
     TIVMap,
 )
 from eitprocessing.plotting.helpers import AbsolutePercentFormatter, AbsoluteScalarFormatter
+from eitprocessing.plotting.pixelmap import (
+    PIXELMAP_PLOT_CONFIG_REGISTRY,
+    PixelMapPlotConfig,
+    reset_pixelmap_plot_config,
+    set_pixelmap_plot_config,
+)
 
 
 def test_init_values():
@@ -80,59 +85,80 @@ def test_update():
     assert pm2.label == "test"
     assert pm2 == pm0.update(values=[[1]], label="test")
 
-    pp0 = PlotParameters()
+    pp0 = PixelMapPlotConfig()
     pp1 = copy.replace(pp0, cmap="foo")
     assert pp1.cmap == "foo"
     assert pp1 == pp0.update(cmap="foo")
 
 
-def test_init_plot_parameters():
+def test_init_plot_config():
     pm0 = PixelMap([[0]])
-    pm1 = PixelMap([[0]], plot_parameters={})
+    pm1 = PixelMap([[0]], plot_config={})
 
-    assert pm0.plot_parameters == pm1.plot_parameters
-    assert pm0.plot_parameters is not pm1.plot_parameters
+    assert pm0._plot_config is pm0.plotting.config
+    assert pm1._plot_config is pm1.plotting.config
+    assert isinstance(pm0.plotting.config, PixelMapPlotConfig)
+    assert isinstance(pm1.plotting.config, PixelMapPlotConfig)
 
-    pm2 = PixelMap([[0]], plot_parameters=PlotParameters(facecolor="white"))
-    pm3 = PixelMap([[0]], plot_parameters={"facecolor": "white"})
+    assert pm0.plotting.config == pm1.plotting.config
+    assert pm0.plotting.config is not pm1.plotting.config
 
-    assert pm2.plot_parameters != pm0.plot_parameters
-    assert pm2.plot_parameters == pm3.plot_parameters
+    pm2 = PixelMap([[0]], plot_config=PixelMapPlotConfig(facecolor="white"))
+    pm3 = PixelMap([[0]], plot_config={"facecolor": "white"})
 
-    pm4 = pm0.update(plot_parameters=PlotParameters(facecolor="white"))
-    pm5 = pm0.update(plot_parameters={"facecolor": "white"})
+    assert pm2.plotting.config != pm0.plotting.config
+    assert pm2.plotting.config == pm3.plotting.config
 
-    assert pm4.plot_parameters != pm0.plot_parameters
-    assert pm4.plot_parameters == pm2.plot_parameters
-    assert pm4.plot_parameters == pm5.plot_parameters
+    with pytest.raises(TypeError, match=r"PixelMapPlotConfig.__init__\(\) got an unexpected keyword argument"):
+        _ = PixelMap([[0]], plot_config={"non_existing": None})
 
-    pm6 = pm2.update(plot_parameters={"cmap": "inferno"})
-    pm7 = pm2.update(plot_parameters=PlotParameters(cmap="inferno"))
-    pm8 = pm2.update(plot_parameters=pm2.plot_parameters.update(cmap="inferno"))
-
-    # when passing a dict as plot_parameters, plot_parameters is *updated*, not *replaced*
-    # pm6 works the same as pm8 internally
[…]
