+++ eitprocessing/datahandling/pixelmap.py
@@ -11,126 +11,57 @@ consists of:
 - PendelluftMap: Specialized map for positive-only pendelluft values (severity, no direction).
 - SignedPendelluftMap: Specialized map for signed pendelluft values (severity and direction/phase).
 
-Each map type provides appropriate default colormaps and normalizations suitable for
-their specific use case, along with methods for visualization and data manipulation.
-
-Plotting parameters for each map are managed via a `PlotParameters` dataclass, which allows for flexible configuration
-of colormap, normalization, colorbar, and other display options. These can be customized per map type or per instance.
+Plotting configurations for each map are managed via a class inheriting from `PixelMapPlotConfig`, which allows for
+flexible configuration of colormap, normalization, colorbar, and other display options. These can be customized per map
+type or per instance. `PixelMapPlotConfig` is found in `eitprocessing.plotting.pixelmap`.
 
 All classes are immutable to ensure data integrity during analysis pipelines.
 
-Both `PixelMap` and `PlotParameters` provide a `replace` method, which allows you to create a copy of the object with
-one or more attributes replaced, similar to `dataclasses.replace()`. This is useful for updating data or configuration
-in an immutable and chainable way, and supports partial updates of nested configuration (e.g., updating only a single
-plotting parameter).
+`PixelMap` provides a `replace` method, which allows you to create a copy of the object with one or more attributes
+replaced, similar to `dataclasses.replace()` (or `copy.replace() in Python ≥3.13). This is useful for updating data or
+configuration in an immutable and chainable way, and supports partial updates of nested configuration (e.g., updating
+only a single plotting configuration).
+
+Mathematical Operations:
+    `PixelMap` instances support basic mathematical operations (+, -, *, /) with other `PixelMap` instances, arrays, or
+    scalar values. The operations are applied element-wise to the underlying values.
+
+    - Addition (+): Returns a `PixelMap` with values added element-wise.
+
+    - Subtraction (-): Returns a `DifferenceMap` with values subtracted element-wise.
+
+    - Multiplication (*): Returns a `PixelMap` with values multiplied element-wise.
+
+    - Division (/): Returns a Pi`xelMap with values divided element-wise. Division by zero results in NaN values with a
+      warning.
+
+    When operating with another `PixelMap` of any type, operations typically return the base PixelMap type, except for
+    subtraction which returns a DifferenceMap. When operating with scalars or arrays, operations return the same type as
+    the original `PixelMap`.
+
+    Note: Some `PixelMap` subclasses (like `TIVMap` and `PerfusionMap`) do not allow negative values. Operations that
+    might produce negative values with these maps will display appropriate warnings.
 """
 
 from __future__ import annotations
 
 import sys
 import warnings
-from copy import deepcopy
-from dataclasses import KW_ONLY, MISSING, asdict, dataclass, field, replace
+from dataclasses import KW_ONLY, InitVar, asdict, dataclass, field, replace
 from typing import TYPE_CHECKING, ClassVar, Literal, TypeVar, cast
 
-import matplotlib as mpl
 import numpy as np
-from frozendict import frozendict
-from matplotlib import colorbar
-from matplotlib import pyplot as plt
-from matplotlib.colors import CenteredNorm, Colormap, LinearSegmentedColormap, Normalize
-from matplotlib.ticker import PercentFormatter
 from numpy import typing as npt
 from typing_extensions import Self
-
-from eitprocessing.plotting.helpers import AbsolutePercentFormatter, AbsoluteScalarFormatter
 
 if TYPE_CHECKING:
[…]

[…]
