### 2025-07-31 701cc93
Move config to own module
