+++ eitprocessing/plotting/pixelmap.py
@@ -28,8 +28,7 @@ Examples:
 """
 
 from copy import deepcopy
-from dataclasses import MISSING, Field, dataclass, field, fields, replace
-from typing import Self, TypeVar, get_type_hints
+from dataclasses import dataclass, field
 
 import matplotlib as mpl
 import numpy as np
@@ -41,6 +40,7 @@ from matplotlib.colors import CenteredNorm, Colormap, LinearSegmentedColormap, N
 from matplotlib.image import AxesImage
 from matplotlib.ticker import PercentFormatter
 
+from eitprocessing.config import Config
 from eitprocessing.datahandling.pixelmap import (
     DifferenceMap,
     ODCLMap,
@@ -53,50 +53,6 @@ from eitprocessing.datahandling.pixelmap import (
 from eitprocessing.plotting.helpers import AbsolutePercentFormatter, AbsoluteScalarFormatter
 
 ColorType = str | tuple[float, float, float] | tuple[float, float, float, float] | float | Colormap
-
-
-T = TypeVar("T")
-
-
-def _get_field_type(field_: Field[T], cls: type) -> type[T]:
-    # If using future annotations, resolve string to real type
-    type_hints = get_type_hints(cls)
-    return type_hints[field_.name]
-
-
-@dataclass(frozen=True, kw_only=True)
-class Config:
-    """Base class for configuration."""
-
-    def __post_init__(self):
-        for field_ in fields(self):
-            if _get_field_type(field_, self.__class__) in (dict, frozendict):
-                # Convert dict fields to frozendict for immutability
-                default_factory = field_.default_factory
-                default_value = default_factory() if default_factory is not MISSING else {}
-                merged = default_value | (getattr(self, field_.name) or {})
-                object.__setattr__(self, field_.name, frozendict(merged))
-
-    def __replace__(self, /, **changes) -> Self:
-        """Return a copy of the of the Config instance replacing attributes.
-
-        Similar to dataclass.replace(), but with special handling of `colorbar_kwargs`. `colorbar_kwargs` is updated
-        with the provided dictionary, rather than replaced.
-
-        Args:
-            **changes: New values for attributes to replace.
-
-        Returns:
-            Self: A new instance with the replaced attributes.
-        """
-        for field_ in fields(self):
-            if _get_field_type(field_, self.__class__) in (dict, frozendict) and field_.name in changes:
-                # Instead of replacing the existing with the new dict, merge the changes
-                changes[field_.name] = getattr(self, field_.name) | changes[field_.name]
-
-        return replace(self, **changes)
-
-    update = __replace__
 
 
 def _get_zero_norm() -> Normalize:
