### 2025-08-01 52e6a38
Move plot config registry
