### 2025-08-01 187298e
Fix typing and Python version issues
