### 2025-08-01 0d16ada
Fix naming and documentation based on suggestions by JF
