### 2025-08-01 abf0ee0
Fix for Python 3.10 compatibility issue
