### 2025-08-01 f50583c
Add warning when non-boolean 0-values are converted to NaN
