### 2025-08-01 e566d12
Ignore version-specific code in coverage reports
