+++ pyproject.toml
@@ -91,7 +91,10 @@ command_line = "-m pytest"
 omit = ["tests/*", "*/tests/*"]
 
 [tool.coverage.report]
-exclude_lines = ["if TYPE_CHECKING:"]
+exclude_lines = [
+    "if TYPE_CHECKING:",
+    "if sys\\.version_info"
+]
 
 [tool.setuptools.packages.find]
 include = ["eitprocessing*", "eitprocessing.*"]
