### 2025-08-01 801211a
Reduce redundancy
