### 2025-08-01 65460dd
Replace from_mean with from_aggregate
