+++ eitprocessing/plotting/pixelmap.py
@@ -319,7 +319,7 @@ class PendelluftMapPlotConfig(PixelMapPlotConfig):
     """
 
     cmap: str | Colormap = field(
-        default_factory=lambda: LinearSegmentedColormap.from_list("Perfusion", ["black", "forestgreen"])
+        default_factory=lambda: LinearSegmentedColormap.from_list("Pendelluft", ["black", "forestgreen"])
     )
     norm: Normalize = field(default_factory=_get_zero_norm)
     colorbar_kwargs: frozendict = field(default_factory=lambda: frozendict(label="Pendelluft"))
@@ -338,7 +338,9 @@ class SignedPendelluftMapPlotConfig(PixelMapPlotConfig):
     """
 
     cmap: str | Colormap = field(
-        default_factory=lambda: LinearSegmentedColormap.from_list("Perfusion", ["deeppink", "black", "forestgreen"])
+        default_factory=lambda: LinearSegmentedColormap.from_list(
+            "SignedPendelluft", ["deeppink", "black", "forestgreen"]
+        )
     )
     colorbar_kwargs: frozendict = field(default_factory=lambda: frozendict(label="Pendelluft"))
     absolute: bool = True
