### 2025-08-01 fe07fc6
Re-add improved docstring descriptions for subclasses of pixelmap
