### 2025-08-04 f604459
Add update function to PixelMask
