### 2025-08-04 e76c484
Add label to PixelMask class
