### 2025-08-04 13e9a59
Add warning for all-NaN masks including suppress boolean
