### 2025-08-04 e7e3e36
Update pixelmask tests to suppress warnings
