+++ tests/test_pixelmask.py
@@ -35,7 +35,7 @@ def test_pixelmask_init_with_boolean_array():
 
 def test_pixelmask_init_with_integer_array():
     values = np.random.default_rng().integers(0, 2, (10, 10))
-    mask = PixelMask(values)
+    mask = PixelMask(values, suppress_zero_conversion_warning=True)
 
     assert np.all((np.isnan(mask.mask)) | (mask.mask == 1.0))
     assert np.array_equal(np.isnan(mask.mask), values == 0)
@@ -46,7 +46,7 @@ def test_pixelmask_init_with_float_array_non_weighted():
     values = np.random.default_rng().random((10, 10)).round()
     assert np.all((values == 0.0) | (values == 1.0))
 
-    mask = PixelMask(values)
+    mask = PixelMask(values, suppress_zero_conversion_warning=True)
 
     assert np.all((np.isnan(mask.mask)) | (mask.mask == 1.0))
     assert np.array_equal(np.isnan(mask.mask), values == 0.0)
@@ -57,7 +57,7 @@ def test_pixelmask_init_with_float_array_weighted():
     values = np.random.default_rng().random((10, 10))
     values[values < 0.5] = 0.0
 
-    mask = PixelMask(values)
+    mask = PixelMask(values, suppress_zero_conversion_warning=True)
 
     assert np.array_equal(np.isnan(mask.mask), values < 0.5)  # NaN for values < 0.5
     assert np.array_equal(~np.isnan(mask.mask), values >= 0.5)  # non-NaN for values >= 0.5
@@ -66,7 +66,7 @@ def test_pixelmask_init_with_float_array_weighted():
 
 def test_pixelmask_init_with_list():
     values = [[1, 0, 1], [0, 1, 0]]
-    mask = PixelMask(values)
+    mask = PixelMask(values, suppress_zero_conversion_warning=True)
 
     assert np.array_equal(mask.mask, np.array([[1.0, np.nan, 1.0], [np.nan, 1.0, np.nan]]), equal_nan=True)
 
@@ -142,12 +142,12 @@ def test_pixelmask_is_weighted_true():
 
 
 def test_pixelmask_is_weighted_false():
-    pm = PixelMask(np.round(np.random.default_rng().random((10, 10))))
+    pm = PixelMask(np.round(np.random.default_rng().random((10, 10))), suppress_zero_conversion_warning=True)
     assert not pm.is_weighted
 
 
 def test_pixelmask_apply_numpy_array():
-    pm = PixelMask([[0, 1], [1, 0]])
+    pm = PixelMask([[0, 1], [1, 0]], suppress_zero_conversion_warning=True)
     data = np.array([[1, 2], [3, 4]])
     result = pm.apply(data)
     assert np.array_equal(result, np.array([[np.nan, 2], [3, np.nan]]), equal_nan=True)
@@ -159,7 +159,7 @@ def test_pixelmask_apply_numpy_array():
 
 
 def test_pixelmask_apply_numpy_array_higher_dimensions():
-    pm = PixelMask([[0, 1], [1, 0]])
+    pm = PixelMask([[0, 1], [1, 0]], suppress_zero_conversion_warning=True)
     data = np.array([[[1, 2], [3, 4]], [[5, 6], [7, 8]]])
     result = pm.apply(data)
     assert np.array_equal(result, np.array([[[np.nan, 2], [3, np.nan]], [[np.nan, 6], [7, np.nan]]]), equal_nan=True)
@@ -167,7 +167,9 @@ def test_pixelmask_apply_numpy_array_higher_dimensions():
 
 def test_pixelmask_apply_eitdata(draeger1: Sequence):
     eit_data = draeger1.eit_data["raw"]
[…]
