### 2025-08-04 aedf2af
Add PixelMaskCollection
