+++ eitprocessing/roi/pixelmaskcollection.py
@@ -1,0 +1,284 @@
+import warnings
+from collections.abc import Hashable
+from dataclasses import dataclass, field, replace
+from typing import TypeVar, cast, overload
+
+import numpy as np
+from frozendict import frozendict
+from typing_extensions import Self
+
+from eitprocessing.datahandling.eitdata import EITData
+from eitprocessing.datahandling.pixelmap import PixelMap
+from eitprocessing.roi import PixelMask
+
+T = TypeVar("T", bound=EITData | PixelMap)
+
+
+@dataclass(frozen=True)
+class PixelMaskCollection:
+    """A collection of pixel masks, each representing a specific region of interest (ROI) in the EIT data.
+
+    This class allows for the application of multiple masks to numpy arrays, EIT data or pixel maps, enabling the
+    extraction of specific regions based on the defined masks.
+
+    PixelMasks can be labelled with strings, or anonymous (`label=None`). Either all or none of the masks should have a
+    label; it is not possible to mix labelled and unlabelled masks.
+
+    At initialization, masks can be provided as a list or a dictionary. If provided as a list, the masks are converted
+    to a dictionary with either their labels as keys (if all masks have labels) or their indices as keys (if none have
+    labels). If provided as a dictionary, the keys must match the labels of the masks; anonymous PixelMaps can not be
+    provided as a dictionary.
+
+    When the `apply` method is called, it applies each mask to the provided data and returns a dictionary
+    mapping each mask's key (label or index) to the resulting masked data.
+
+    Example:
+    >>> mask_collection = PixelMaskCollection(
+            [
+                PixelMask(right_lung_data, label="right lung"),
+                PixelMask(left_lung_data, label="left lung"),
+            ],
+            label="lung masks",
+        )
+
+    Args:
+        masks (dict): A dictionary mapping mask names to their corresponding pixel masks.
+        label (str | None): An optional label for the collection of masks.
+
+    Attributes:
+        is_anonymous (bool):
+            A boolean indicating whether all masks in the collection are anonymous (i.e., have no label).
+
+    """
+
+    masks: frozendict[Hashable, PixelMask]
+    label: str | None = field(default=None)
+
+    def __init__(
+        self,
+        masks: dict[Hashable, PixelMask] | frozendict[Hashable, PixelMask] | list[PixelMask],
+        *,
+        label: str | None = None,
+    ):
+        object.__setattr__(self, "masks", self._validate_and_convert_input_masks(masks))
+        object.__setattr__(self, "label", label)
+
+    def _validate_and_convert_input_masks(
+        self, masks: dict[Hashable, PixelMask] | frozendict[Hashable, PixelMask] | list[PixelMask]
+    ) -> frozendict[Hashable, PixelMask]:
+        """Validate the input masks and convert to a frozendict.
+
+        The input masks are valid if:
+        - The input is a list with all labelled or all anonymous PixelMask instances.
+        - The input is a dictionary where each PixelMask instance's label matches the key.
+        """
+        # Check for type before checking for length, e.g., and empty string should raise a TypeError, not ValueError
+        if not isinstance(masks, (list, dict, frozendict)):
+            msg = f"Expected a list or a dictionary, got {type(masks)}."
+            raise TypeError(msg)
+
+        if not masks:
+            msg = "A PixelMaskCollection should contain at least one mask."
+            raise ValueError(msg)
+
+        if isinstance(masks, list):
+            return self._validate_and_convert_list_input(masks)
+
+        return self._validate_and_convert_dict_input(masks)
+
+    def _validate_and_convert_list_input(self, masks: list[PixelMask]) -> frozendict[Hashable, PixelMask]:
+        if any(not isinstance(mask, PixelMask) for mask in masks):
+            msg = "All items in the collection must be instances of PixelMask."
+            raise TypeError(msg)
[…]
