### 2025-08-04 d39d45f
Add tests for PixelMaskCollection
