+++ tests/test_pixelmask_collection.py
@@ -1,0 +1,450 @@
+from collections.abc import Callable
+
+import numpy as np
+import pytest
+
+from eitprocessing.datahandling.eitdata import EITData
+from eitprocessing.datahandling.pixelmap import PixelMap
+from eitprocessing.datahandling.sequence import Sequence
+from eitprocessing.roi import PixelMask
+from eitprocessing.roi.pixelmaskcollection import PixelMaskCollection
+
+
+@pytest.fixture
+def numpy_array() -> Callable:
+    return lambda: np.random.default_rng().random((32, 32))
+
+
+@pytest.fixture
+def pixel_map(numpy_array: Callable) -> Callable:
+    return lambda: PixelMap(numpy_array(), label="test_map")
+
+
+@pytest.fixture
+def anonymous_boolean_mask(numpy_array: Callable) -> Callable:
+    return lambda: PixelMask(numpy_array() > 0.5)
+
+
+@pytest.fixture
+def labelled_boolean_mask(anonymous_boolean_mask: Callable):
+    return lambda label: anonymous_boolean_mask().update(label=label)
+
+
+@pytest.fixture
+def anonymous_float_mask(numpy_array: Callable) -> Callable:
+    def factory() -> PixelMask:
+        numpy_array_ = numpy_array() * 1.2 - 0.2
+        numpy_array_[numpy_array_ <= 0] = np.nan
+        return PixelMask(numpy_array_)
+
+    return factory
+
+
+@pytest.fixture
+def labelled_float_mask(anonymous_float_mask: Callable) -> Callable:
+    return lambda label: anonymous_float_mask().update(label=label)
+
+
+def test_boolean_mask_factory_unique(anonymous_boolean_mask: Callable):
+    mask1 = anonymous_boolean_mask()
+    mask2 = anonymous_boolean_mask()
+    assert not np.array_equal(mask1.mask, mask2.mask)
+
+
+def test_float_mask_factory_unique(anonymous_float_mask: Callable):
+    mask1 = anonymous_float_mask()
+    mask2 = anonymous_float_mask()
+    assert not np.array_equal(mask1.mask, mask2.mask)
+
+
+def test_init_with_label(anonymous_boolean_mask: Callable):
+    mask = anonymous_boolean_mask()
+    collection = PixelMaskCollection([mask], label="test_collection")
+    assert collection.label == "test_collection"
+
+
+def test_init_without_label(anonymous_boolean_mask: Callable):
+    mask = anonymous_boolean_mask()
+    collection = PixelMaskCollection([mask])
+    assert collection.label is None
+
+
+def test_init_with_labelled_masks(labelled_boolean_mask: Callable):
+    pm1 = labelled_boolean_mask("mask1")
+    pm2 = labelled_boolean_mask("mask2")
+    collection1 = PixelMaskCollection([pm1, pm2])
+    assert collection1.masks == {"mask1": pm1, "mask2": pm2}
+
+    collection2 = PixelMaskCollection({"mask2": pm2, "mask1": pm1})
+    assert collection2 == collection1
+
+
+def test_init_with_anonymous_masks(anonymous_boolean_mask: Callable):
+    pm1 = anonymous_boolean_mask()
+    pm2 = anonymous_boolean_mask()
+    collection1 = PixelMaskCollection([pm1, pm2])
+    assert collection1.masks == {0: pm1, 1: pm2}
+
+    collection2 = PixelMaskCollection({0: pm1, 1: pm2})
+    assert collection1 == collection2
+
+
+def test_init_with_dict_integer_keys_unordered_raises(anonymous_boolean_mask: Callable):
+    pm1 = anonymous_boolean_mask()
[…]
