### 2025-08-04 b6d0346
Fix docstrings
