### 2025-08-05 13d4148
Add combine method
