### 2025-08-05 81172ab
Add PixelMask(Collection) to docs
