+++ docs/api/roi/roi.md
@@ -1,0 +1,4 @@
+::: eitprocessing.roi
+    options:
+        members:
+            - PixelMask
