### 2025-08-07 878772e
Improve rate detection
- add refinement option for estimated frequency
- make welch function faster
- reduce code duplication
- improve documentation
- add captures to extract intermediate data
