+++ eitprocessing/features/rate_detection.py
@@ -1,151 +1,269 @@
-import itertools
+import warnings
 from dataclasses import dataclass
-from typing import Final, Literal
+from typing import TYPE_CHECKING, Final, Literal
 
 import numpy as np
 from scipy import signal
 
 from eitprocessing.datahandling.eitdata import EITData
-from eitprocessing.datahandling.intervaldata import IntervalData
+
+if TYPE_CHECKING:
+    from eitprocessing.plotting.rate_detection import RateDetectionPlotting
 
 MINUTE: Final = 60
+MIN_WELCH_WINDOW_LENGTH: Final = 10
 
 
-@dataclass
+@dataclass(frozen=True)
 class RateDetection:
-    """Detect the respiratory and heart rate from EIT pixel data."""
+    """Detect the respiratory and heart rate from EIT pixel data.
+
+    This algorithm attempts to detect the respiratory and heart rate from EIT pixel data. It is based on the observation
+    that many high-amplitude pixels have the respiratory rate as the main frequency, while in low-amplitude pixels the
+    power of the heart rate is relatively high. The algorithm uses Welch's method to estimate the power spectrum of the
+    summed pixel data and individual pixels. It then identifies the respiratory rate as the frequency with the highest
+    power for the summed pixels within the specified range. The power spectra of the individual pixels are normalized
+    and averaged. The normalized power spectrum of the summed pixels is subtracted from the average of the the
+    normalized individual power spectra. The frequency with the highest relative power in this difference within the
+    specified range is taken as the heart rate.
+
+    If either rate is variable, the algorithm will in most cases return an average frequency. If there are multiple
+    distinct frequencies, e.g., due to a change in the controlled respiratory rate, multiple peaks will be visible in
+    the power spectrum. The algorithm will only return the frequency with the highest power in the specified range.
+
+    The algorithm can't distinguish between the respiratory and heart rate if they are too close together, especially in
+    very short signals (or when the Welch window is short). The algorithm can distinguish between both rates if the
+    heart rate is at one of the harmonics of the respiratory rate.
+
+    If the `refine_estimated_frequency` attribute is set to False, the estimated frequency is simply the location of the
+    peak of the power Welch spectrum. Since Welch's method results in a limited number of frequency bins, this can lead
+    to inaccuracies, especially with short or low-sample frequency data. If `refine_estimated_frequency` is set to True
+    (default), the estimated frequency is refined using parabolic interpolation, which often yields more accurates
+    results, even with short signals, low signal-to-noise ratio and very similar (but distinct) respiratory and heart
+    rates. See e.g. [Quadratic Interpolation of Spectral
+    Peaks](https://ccrma.stanford.edu/~jos/sasp/Quadratic_Interpolation_Spectral_Peaks.html).
+
+    The respiratory and heart rate limits can be set when initializing this algorithms. Defaults values for adults and
+    neonates are set in `DEFAULT_RATE_LIMITS`.
+
+    Although the algorithm might perform reasonably on data as short as 10 seconds (200 samples), it is recommended to
+    use at least 30 seconds of data for reliable results. Longer data wil yield more reliable results, but only if the
+    respiratory and heart rates are stable.
+
+    Note that the algorithm performs best if no large changes in end-expiratory impedance occur in the data, the data is
+    unfiltered, and the respiratory and heart rates are relatively stable.
+
+    Attributes:
+        subject_type: The type of subject, either "adult" or "neonate". This affects the default settings for the
[…]
