### 2025-08-07 41e7a76
Add RateDetectionPlotting
