+++ eitprocessing/plotting/rate_detection.py
@@ -1,0 +1,173 @@
+from typing import overload
+
+import numpy as np
+from matplotlib import pyplot as plt
+from matplotlib.figure import Figure, SubFigure
+from matplotlib.ticker import MaxNLocator
+
+from eitprocessing.features.rate_detection import RateDetection
+
+MINUTE = 60
+
+
+class RateDetectionPlotting:
+    """Utility class to plot the results of the RateDetection algorithm.
+
+    This class can be accessed through the `plotting` attribute of the RateDetection class.
+
+    Example:
+    ```python
+    >>> rd = RateDetection("adult")
+    >>> captures = {}
+    >>> estimated_respiratory_rate, estimated_heart_rate = rd.detect_respiratory_heart_rate(eit_data, captures=captures)
+    >>> fig = rd.plotting(**captures)
+    >>> fig.savefig(...)
+    ```
+    """
+
+    def __init__(self, obj: RateDetection):
+        self.obj = obj
+
+    @overload
+    def plot(
+        self,
+        *,
+        frequencies: np.ndarray,
+        normalized_total_power: np.ndarray,
+        average_normalized_pixel_power: np.ndarray,
+        diff_total_averaged_power: np.ndarray,
+        estimated_respiratory_rate: float,
+        estimated_heart_rate: float,
+        fig: None,
+    ) -> Figure: ...
+
+    @overload
+    def plot(
+        self,
+        *,
+        frequencies: np.ndarray,
+        normalized_total_power: np.ndarray,
+        average_normalized_pixel_power: np.ndarray,
+        diff_total_averaged_power: np.ndarray,
+        estimated_respiratory_rate: float,
+        estimated_heart_rate: float,
+        fig: SubFigure,
+    ) -> SubFigure: ...
+
+    @overload
+    def plot(
+        self,
+        *,
+        frequencies: np.ndarray,
+        normalized_total_power: np.ndarray,
+        average_normalized_pixel_power: np.ndarray,
+        diff_total_averaged_power: np.ndarray,
+        estimated_respiratory_rate: float,
+        estimated_heart_rate: float,
+        fig: Figure,
+    ) -> Figure: ...
+
+    def plot(
+        self,
+        *,
+        frequencies: np.ndarray,
+        normalized_total_power: np.ndarray,
+        average_normalized_pixel_power: np.ndarray,
+        diff_total_averaged_power: np.ndarray,
+        estimated_respiratory_rate: float,
+        estimated_heart_rate: float,
+        fig: Figure | SubFigure | None = None,
+    ) -> Figure | SubFigure:
+        """Plot the results of the RateDetection algorithm.
+
+        This method is intended to be used after running the `detect_respiratory_heart_rate` method of the
+        `RateDetection` class. It visualizes the frequency analysis results, including the normalized power
+        spectrum, average normalized pixel spectrum, and the difference between the two.
+
+        NB: Although the frequencies, rate limits and estimated rates are provided in Hz, they are visualized in
+        breaths/beats per minute (bpm).
+
+        Arguments:
+            frequencies: Frequencies used in the analysis, in Hz.
+            normalized_total_power: Normalized total power spectrum.
+            average_normalized_pixel_power: Average normalized pixel power spectrum.
+            diff_total_averaged_power:
+                Difference between the normalized summed power spectra and the normalized total power.
+            estimated_respiratory_rate: Estimated respiratory rate in Hz.
+            estimated_heart_rate: Estimated heart rate in Hz.
[…]
