### 2025-08-07 85fe650
Add tests for rate detection
