+++ tests/test_rate_detection.py
@@ -1,15 +1,323 @@
 # %%
 
-from eitprocessing.datahandling.loading import load_eit_data
-from eitprocessing.features.rate_detection import RateDetection
+import itertools
+from collections.abc import Callable
+
+import numpy as np
+import pytest
+from scipy import signal
+
+from eitprocessing.datahandling.eitdata import EITData
+from eitprocessing.datahandling.sequence import Sequence
+from eitprocessing.features.rate_detection import DEFAULT_RATE_LIMITS, RateDetection
+from eitprocessing.plotting.rate_detection import RateDetectionPlotting
+
+MINUTE = 60
 
 
-def test_ratedetection_works():
-    path = "tests/test_data/Draeger_Test.bin"
-    sequence = load_eit_data(path, vendor="draeger")
-    subsequence = sequence.t[56600:56800]
+@pytest.fixture
+def signal_factory() -> Callable[..., EITData]:
+    """Create a factory function to generate synthetic EIT data with specified characteristics.
 
-    rd = RateDetection(subject_type="adult")
-    rr, hr = rd.detect_respiratory_heart_rate(subsequence.eit_data["raw"])
-    assert rr.values[0] == 25.0 / 60
-    assert hr.values[0] == 71.0 / 60
+    The generated EIT data is suitable for testing the RateDetection algorithm. It includes high and low power frequency
+    components, noise, and allows for customization of parameters such as duration, sample frequency, shape, and
+    amplitudes. The high and low power signals are generated using triangular waveforms.
+
+    The high power signal has high power in half the pixels, and lower power in the other half. This simulates a limited
+    number of pixels with mainly respiratory signal. The low power signal has lower power in all pixels, simulating a
+    more uniform distribution of heart rate signal. In most data, a single high power frequency and single low power
+    frequency should be used, but the factory allows for multiple frequencies to be specified to simulate shifting or
+    irregular signals. The exact amplitude of each frequency component is randomized to add variability to the generated
+    data.
+
+    The frequencies can be increased or decreased from the start to the end of the signal. The frequency at the end of
+    the signal is the base frequency multiplied by the frequency scale factor. A frequency scale factor of 1.0 will
+    result in a constant frequency, values greater than 1.0 will lead to an increasing frequency, and values less than
+    1.0 will lead to a decreasing frequency.
+
+    The first row of the pixel impedance is set to NaN to simulatte missing data.
+
+    Arguments:
+        high_power_frequencies: Frequencies for the high power signal in Hz.
+        low_power_frequencies: Frequencies for the low power signal in Hz.
+        duration: Duration of the signal in seconds (default is 60.0).
+        sample_frequency: Sample frequency in Hz (default is 20.0).
+        shape: Shape of the pixel impedance array (default is (32, 32)).
+        low_power_amplitude: Amplitude of the low power signal (default is 0.2).
+        noise_amplitude: Amplitude of the noise added to the signal (default is 0.01).
+        captures: Optional dictionary to capture generated signals and time for testing purposes.
+    """
+
+    def factory(  # noqa: PLR0913
+        *,
+        high_power_frequencies: tuple[float, ...],
[…]
