### 2025-08-07 9f113d1
Ignore captures in coverage reporting
