+++ pyproject.toml
@@ -94,7 +94,8 @@ omit = ["tests/*", "*/tests/*"]
 exclude_lines = [
     "if TYPE_CHECKING:",
     "if sys\\.version_info",
-    "@overload"
+    "@overload",
+    "if captures is not None:"
 ]
 
 [tool.setuptools.packages.find]
