### 2025-08-07 bc6ae94
Rename detect_respiratory_heart_rate to apply
