### 2025-08-07 585fbe9
Add more captures in signal_factory
