### 2025-08-07 c431201
Reduce default Welch window size
