### 2025-08-07 3e7a468
Create RateDetection documentation and move Features to separate pages
