+++ docs/api/features/moving_average.md
@@ -1,0 +1,1 @@
+::: eitprocessing.features.moving_average.MovingAverage
