+++ mkdocs.yml
@@ -59,11 +59,15 @@ nav:
         - Breath: api/datacontainers/breath.md
         - Event: api/datacontainers/event.md
         - DataCollection: api/datacontainers/datacollection.md
+      - Features:
+        - Breath Detection: api/features/breath_detection.md
+        - Pixel Breath: api/features/pixel_breath.md
+        - Rate Detection: api/features/rate_detection.md
+        - Moving Average: api/features/moving_average.md
       - api/filters.md
       - Regions of Interest:
         - Regions of Interest: api/roi/roi.md
         - PixelMaskCollection: api/roi/pixelmaskcollection.md
-      - api/features.md
       - api/parameters.md
       - api/categories.md
       - api/mixins.md
