### 2025-08-07 ad27aa7
Add test for warning
