### 2025-08-07 f6f06b3
Fix erroneous value check
