### 2025-08-07 fadc849
Remove cases further from edges of range
