+++ .github/workflows/test_build_documentation.yml
@@ -5,6 +5,7 @@ on:
     branches:
       - main
       - develop
+      - feature/211_rr_hr_detection_psomhorst
   pull_request:
     types: [opened, synchronize, reopened, ready_for_review]
     branches:
@@ -31,6 +32,11 @@ jobs:
         run: |
           ln -s ../../notebooks docs/examples
           mkdocs build
+      - name: Upload built docs
+        uses: actions/upload-artifact@v4
+        with:
+          name: mkdocs-site
+          path: site/
       - name: Deploy documentation
         if: github.ref == 'refs/heads/main'
         run: mkdocs gh-deploy
