### 2025-08-07 9f184a7
Revert "Temporarily build docs for download"
This reverts commit f4b08d715e947faf6d6d6824414fb558ffd3fd71.
