### 2025-08-07 6490606
Add notebook to documentation
