+++ .github/workflows/test_build_documentation.yml
@@ -28,7 +28,9 @@ jobs:
           extract-data: false
           python-version: "3.10"
       - name: Build documentation
-        run: mkdocs build
+        run: |
+          ln -s ../../notebooks docs/examples
+          mkdocs build
       - name: Deploy documentation
         if: github.ref == 'refs/heads/main'
         run: mkdocs gh-deploy
