+++ pyproject.toml
@@ -45,6 +45,7 @@ dependencies = [
     "typing_extensions",
     "pyyaml",
     "frozendict>=2.4.6",
+    "mkdocs-jupyter>=0.25.1",
 ]
 
 [project.optional-dependencies]
