### 2025-08-08 18c2f9a
Update docstring about input masks
