### 2025-08-08 632035b
Reduce complexity
