### 2025-08-08 c189b06
Add test for mixed dict
