### 2025-08-08 79475c5
Update error message in tests
