### 2025-08-08 753020e
Add estimated rate annotations to plot
