### 2025-08-08 8d4938f
Improve docstrings
