### 2025-08-08 5bf3852
Reduce test strictness
