### 2025-08-08 90b7c40
Create first draft for ROILabeller object
