### 2025-08-11 a4ecea0
Return PixelMask and rename class
