### 2025-08-11 d10ae23
Add docstring
