### 2025-08-11 95d0fc8
Update for more flexibility with connectivity
