### 2025-08-11 6151157
Update for more flexibility with connectivity
