+++ tests/test_roi_selector.py
@@ -12,77 +12,65 @@ def make_pixel_mask(array: np.ndarray):
 
 
 def test_basic_region_selection():
-    # Simple 5x5 mask with one large connected region
     arr = np.full((5, 5), np.nan)
-    arr[1:4, 1:4] = True  # 3x3 region
+    arr[1:4, 1:4] = True
     mask = make_pixel_mask(arr)
-
-    selector = ROISelector(min_pixels=5)
+    selector = ROISelector(min_pixels=5, structure="4-connectivity")
     result = selector.select_regions(mask)
-
-    # The combined mask should have the same shape and True in the same places
     expected_mask = np.full(arr.shape, np.nan)
     expected_mask[1:4, 1:4] = 1
     np.testing.assert_array_equal(result.mask, expected_mask)
 
 
 def test_region_smaller_than_threshold_is_excluded():
-    # One small region (2 pixels) and one large region (6 pixels)
     arr = np.full((5, 5), np.nan)
-    arr[0, 0:2] = True  # small region
-    arr[2:4, 2:5] = True  # large region
+    arr[0, 0:2] = True
+    arr[2:4, 2:5] = True
     mask = make_pixel_mask(arr)
-
-    selector = ROISelector(min_pixels=5)
+    selector = ROISelector(min_pixels=5, structure="4-connectivity")
     result = selector.select_regions(mask)
-
-    # Small region should be excluded
     expected = np.zeros_like(arr, dtype=float)
     expected[2:4, 2:5] = 1
     expected[expected == 0] = np.nan
-
     np.testing.assert_array_equal(result.mask, expected)
 
 
 def test_no_regions_above_threshold_warns_and_returns_empty():
-    # Mask has a single small region below threshold
     arr = np.full((4, 4), np.nan)
-    arr[0, 0] = True  # 1 pixel only
+    arr[0, 0] = True
     mask = make_pixel_mask(arr)
-
-    selector = ROISelector(min_pixels=5)
+    selector = ROISelector(min_pixels=5, structure="4-connectivity")
     with pytest.warns(UserWarning, match="No regions found above min_pixels threshold."):
         result = selector.select_regions(mask)
-
     assert isinstance(result, PixelMask)
-    # All NaN means no region kept
     assert np.all(np.isnan(result.mask))
 
 
 def test_custom_connectivity():
-    # Two diagonal pixels — only connected if using full connectivity
     arr = np.full((3, 3), np.nan)
     arr[0, 0] = True
     arr[1, 1] = True
     mask = make_pixel_mask(arr)
-
     # Default (4-connectivity) — should warn and return empty
-    selector_default = ROISelector(min_pixels=2)
+    selector_default = ROISelector(min_pixels=2, structure="4-connectivity")
     with pytest.warns(UserWarning, match="No regions found above min_pixels threshold."):
         result_default = selector_default.select_regions(mask)
     assert np.all(np.isnan(result_default.mask))
-
     # 8-connectivity — diagonal pixels connected
-    structure = generate_binary_structure(2, 2)  # 8-connected
-    selector_diag = ROISelector(min_pixels=2, structure=structure)
+    selector_diag = ROISelector(min_pixels=2, structure="8-connectivity")
     result_diag = selector_diag.select_regions(mask)
     assert not np.all(np.isnan(result_diag.mask))
+    # Custom structure
+    custom_structure = generate_binary_structure(2, 2)
+    selector_custom = ROISelector(min_pixels=2, structure=custom_structure)
+    result_custom = selector_custom.select_regions(mask)
+    assert not np.all(np.isnan(result_custom.mask))
[…]
