### 2025-08-11 7e7bc9a
Update parameter name for minimum ROI size
Co-authored-by: Peter Somhorst <127968566+psomhorst@users.noreply.github.com>
