+++ tests/test_roi_size_filter.py
@@ -1,192 +1,166 @@
 import numpy as np
 import pytest
 from scipy.ndimage import generate_binary_structure
-from scipy.ndimage import label as nd_label
 
 from eitprocessing.roi import PixelMask
 from eitprocessing.roi.roi_size_filter import FilterROIBySize
 
 
+def make_pixel_mask(array: np.ndarray):
+    """Helper to wrap numpy array into PixelMask with NaNs for False."""
+    return PixelMask(array, keep_zeros=True)
+
+
 def test_basic_region_selection():
     arr = np.full((5, 5), np.nan)
     arr[1:4, 1:4] = True
-    mask = PixelMask(arr, keep_zeros=True)
-    selector = FilterROIBySize(min_region_size=5, connectivity="1-connectivity")
-    result = selector.apply(mask)
+    mask = make_pixel_mask(arr)
+    selector = FilterROIBySize(min_region_size=5, structure="4-connectivity")
+    result = selector.select_regions(mask)
     expected_mask = np.full(arr.shape, np.nan)
     expected_mask[1:4, 1:4] = 1
-    assert np.array_equal(result.mask, expected_mask, equal_nan=True)
+    np.testing.assert_array_equal(result.mask, expected_mask)
 
 
 def test_region_smaller_than_threshold_is_excluded():
     arr = np.full((5, 5), np.nan)
     arr[0, 0:2] = True
     arr[2:4, 2:5] = True
-    mask = PixelMask(arr, keep_zeros=True)
-    selector = FilterROIBySize(min_region_size=5, connectivity="1-connectivity")
-    result = selector.apply(mask)
+    mask = make_pixel_mask(arr)
+    selector = FilterROIBySize(min_region_size=5, structure="4-connectivity")
+    result = selector.select_regions(mask)
     expected = np.zeros_like(arr, dtype=float)
     expected[2:4, 2:5] = 1
     expected[expected == 0] = np.nan
-    assert np.array_equal(result.mask, expected, equal_nan=True)
+    np.testing.assert_array_equal(result.mask, expected)
 
 
-def test_no_regions_above_threshold_raises():
+def test_no_regions_above_threshold_warns_and_returns_empty():
     arr = np.full((4, 4), np.nan)
     arr[0, 0] = True
-    mask = PixelMask(arr, keep_zeros=True)
-    selector = FilterROIBySize(min_region_size=5, connectivity="1-connectivity")
-    with pytest.raises(RuntimeError, match="No regions found above min_region_size threshold"):
-        selector.apply(mask)
+    mask = make_pixel_mask(arr)
+    selector = FilterROIBySize(min_region_size=5, structure="4-connectivity")
+    with pytest.warns(UserWarning, match="No regions found above min_pixels threshold."):
+        result = selector.select_regions(mask)
+    assert isinstance(result, PixelMask)
+    assert np.all(np.isnan(result.mask))
 
 
 def test_custom_connectivity():
     arr = np.full((3, 3), np.nan)
     arr[0, 0] = True
     arr[1, 1] = True
-    mask = PixelMask(arr, keep_zeros=True)
-
-    # Default (1-connectivity) — should raise
-    selector_default = FilterROIBySize(min_region_size=2, connectivity="1-connectivity")
-    with pytest.raises(RuntimeError, match="No regions found above min_region_size threshold"):
-        selector_default.apply(mask)
-
-    # 2-connectivity — diagonal pixels connected
-    selector_diag = FilterROIBySize(min_region_size=2, connectivity="2-connectivity")
-    result_diag = selector_diag.apply(mask)
+    mask = make_pixel_mask(arr)
+    # Default (4-connectivity) — should warn and return empty
[…]
