### 2025-08-11 20dee08
Update connectivity arguments and documentation
