+++ tests/test_roi_size_filter.py
@@ -15,7 +15,7 @@ def test_basic_region_selection():
     arr = np.full((5, 5), np.nan)
     arr[1:4, 1:4] = True
     mask = make_pixel_mask(arr)
-    selector = FilterROIBySize(min_region_size=5, structure="4-connectivity")
+    selector = FilterROIBySize(min_region_size=5, connectivity="4-connectivity")
     result = selector.select_regions(mask)
     expected_mask = np.full(arr.shape, np.nan)
     expected_mask[1:4, 1:4] = 1
@@ -27,7 +27,7 @@ def test_region_smaller_than_threshold_is_excluded():
     arr[0, 0:2] = True
     arr[2:4, 2:5] = True
     mask = make_pixel_mask(arr)
-    selector = FilterROIBySize(min_region_size=5, structure="4-connectivity")
+    selector = FilterROIBySize(min_region_size=5, connectivity="4-connectivity")
     result = selector.select_regions(mask)
     expected = np.zeros_like(arr, dtype=float)
     expected[2:4, 2:5] = 1
@@ -39,7 +39,7 @@ def test_no_regions_above_threshold_warns_and_returns_empty():
     arr = np.full((4, 4), np.nan)
     arr[0, 0] = True
     mask = make_pixel_mask(arr)
-    selector = FilterROIBySize(min_region_size=5, structure="4-connectivity")
+    selector = FilterROIBySize(min_region_size=5, connectivity="4-connectivity")
     with pytest.warns(UserWarning, match="No regions found above min_pixels threshold."):
         result = selector.select_regions(mask)
     assert isinstance(result, PixelMask)
@@ -52,17 +52,17 @@ def test_custom_connectivity():
     arr[1, 1] = True
     mask = make_pixel_mask(arr)
     # Default (4-connectivity) — should warn and return empty
-    selector_default = FilterROIBySize(min_region_size=2, structure="4-connectivity")
+    selector_default = FilterROIBySize(min_region_size=2, connectivity="4-connectivity")
     with pytest.warns(UserWarning, match="No regions found above min_pixels threshold."):
         result_default = selector_default.select_regions(mask)
     assert np.all(np.isnan(result_default.mask))
     # 8-connectivity — diagonal pixels connected
-    selector_diag = FilterROIBySize(min_region_size=2, structure="8-connectivity")
+    selector_diag = FilterROIBySize(min_region_size=2, connectivity="8-connectivity")
     result_diag = selector_diag.select_regions(mask)
     assert not np.all(np.isnan(result_diag.mask))
     # Custom structure
     custom_structure = generate_binary_structure(2, 2)
-    selector_custom = FilterROIBySize(min_region_size=2, structure=custom_structure)
+    selector_custom = FilterROIBySize(min_region_size=2, connectivity=custom_structure)
     result_custom = selector_custom.select_regions(mask)
     assert not np.all(np.isnan(result_custom.mask))
 
@@ -70,7 +70,7 @@ def test_custom_connectivity():
 def test_empty_mask_returns_empty():
     arr = np.full((4, 4), np.nan)
     mask = make_pixel_mask(arr)
-    selector = FilterROIBySize(min_region_size=1, structure="4-connectivity")
+    selector = FilterROIBySize(min_region_size=1, connectivity="4-connectivity")
     result = selector.select_regions(mask)
     assert np.all(np.isnan(result.mask))
 
@@ -78,7 +78,7 @@ def test_empty_mask_returns_empty():
 def test_all_nan_mask_returns_empty():
     arr = np.full((4, 4), np.nan)
     mask = PixelMask(arr)
[…]
