### 2025-08-11 58b01da
Change to frozen dataclass and update connectivity parsing
