### 2025-08-11 a9ba31e
Change warning to exception
