### 2025-08-11 9b75147
Improve apply method docstring
