### 2025-08-11 4a17af0
Add FilterROIBySize to documentation.
