+++ docs/api/roi/filterbysize.md
@@ -1,0 +1,1 @@
+::: eitprocessing.roi.roi_size_filter
