### 2025-08-11 b7a280d
Update docstring
Co-authored-by: Peter Somhorst <127968566+psomhorst@users.noreply.github.com>
