### 2025-08-11 e8c97c9
Update connectivity inputs
