+++ tests/test_roi_size_filter.py
@@ -11,7 +11,7 @@ def test_basic_region_selection():
     arr = np.full((5, 5), np.nan)
     arr[1:4, 1:4] = True
     mask = PixelMask(arr, keep_zeros=True)
-    selector = FilterROIBySize(min_region_size=5, connectivity="1-connectivity")
+    selector = FilterROIBySize(min_region_size=5, connectivity=1)
     result = selector.apply(mask)
     expected_mask = np.full(arr.shape, np.nan)
     expected_mask[1:4, 1:4] = 1
@@ -23,7 +23,7 @@ def test_region_smaller_than_threshold_is_excluded():
     arr[0, 0:2] = True
     arr[2:4, 2:5] = True
     mask = PixelMask(arr, keep_zeros=True)
-    selector = FilterROIBySize(min_region_size=5, connectivity="1-connectivity")
+    selector = FilterROIBySize(min_region_size=5, connectivity=1)
     result = selector.apply(mask)
     expected = np.zeros_like(arr, dtype=float)
     expected[2:4, 2:5] = 1
@@ -35,7 +35,7 @@ def test_no_regions_above_threshold_raises():
     arr = np.full((4, 4), np.nan)
     arr[0, 0] = True
     mask = PixelMask(arr, keep_zeros=True)
-    selector = FilterROIBySize(min_region_size=5, connectivity="1-connectivity")
+    selector = FilterROIBySize(min_region_size=5, connectivity=1)
     with pytest.raises(RuntimeError, match="No regions found above min_region_size threshold"):
         selector.apply(mask)
 
@@ -47,12 +47,12 @@ def test_custom_connectivity():
     mask = PixelMask(arr, keep_zeros=True)
 
     # Default (1-connectivity) — should raise
-    selector_default = FilterROIBySize(min_region_size=2, connectivity="1-connectivity")
+    selector_default = FilterROIBySize(min_region_size=2, connectivity=1)
     with pytest.raises(RuntimeError, match="No regions found above min_region_size threshold"):
         selector_default.apply(mask)
 
     # 2-connectivity — diagonal pixels connected
-    selector_diag = FilterROIBySize(min_region_size=2, connectivity="2-connectivity")
+    selector_diag = FilterROIBySize(min_region_size=2, connectivity=2)
     result_diag = selector_diag.apply(mask)
     assert not np.all(np.isnan(result_diag.mask))
 
@@ -66,7 +66,7 @@ def test_custom_connectivity():
 def test_empty_mask_raises():
     arr = np.full((4, 4), np.nan)
     mask = PixelMask(arr, keep_zeros=True)
-    selector = FilterROIBySize(min_region_size=1, connectivity="1-connectivity")
+    selector = FilterROIBySize(min_region_size=1, connectivity=1)
     with pytest.raises(RuntimeError, match="No regions found above min_region_size threshold"):
         selector.apply(mask)
 
@@ -74,7 +74,7 @@ def test_empty_mask_raises():
 def test_zeros_are_regions_nans_are_excluded():
     arr = np.array([[0, 1, np.nan], [0, 0.5, np.nan], [np.nan, np.nan, 0]], dtype=float)
     mask = PixelMask(arr, keep_zeros=True)
-    selector = FilterROIBySize(min_region_size=3, connectivity="1-connectivity")
+    selector = FilterROIBySize(min_region_size=3, connectivity=1)
     result_mask = selector.apply(mask).mask
     expected_mask = np.array([[1, 1, np.nan], [1, 1, np.nan], [np.nan, np.nan, np.nan]], dtype=float)
     assert np.array_equal(result_mask, expected_mask, equal_nan=True)
@@ -85,7 +85,7 @@ def test_multiple_large_regions_are_all_included():
     arr[0:2, 0:2] = True
     arr[4:6, 4:6] = True
[…]
