### 2025-08-11 a4841ce
add default min_region_size
