### 2025-08-11 9e13a67
Update scipy.ndimage import
