+++ docs/api/roi/filterbysize.md
@@ -1,1 +1,1 @@
-::: eitprocessing.roi.roi_size_filter
+::: eitprocessing.roi.filter_by_size
