### 2025-08-11 58d7997
Rename apply_filter to apply in ButterworthFilter
