+++ eitprocessing/filters/__init__.py
@@ -1,6 +1,12 @@
 from abc import ABC, abstractmethod
+from typing import TypeVar
 
 import numpy as np
+
+from eitprocessing.datahandling.eitdata import EITData
+from tests.test_breath_detection import ContinuousData
+
+T = TypeVar("T", bound=np.ndarray | ContinuousData | EITData)
 
 
 class TimeDomainFilter(ABC):
@@ -9,6 +15,6 @@ class TimeDomainFilter(ABC):
     available_in_gui = True
 
     @abstractmethod
-    def apply_filter(self, input_data: np.ndarray) -> np.ndarray:
+    def apply(self, input_data: T, **kwargs) -> T:
         """Apply the filter to the input data."""
         ...
