### 2025-08-11 bf9c9c3
Ignore some notebooks/py-files from mkdocs-jupyter
