+++ eitprocessing/filters/mdn.py
@@ -1,20 +1,32 @@
+import copy
+import math
 import warnings
 from dataclasses import dataclass
+from typing import TypeVar, cast, overload
 
 import numpy as np
 from scipy import signal
 
+from eitprocessing.datahandling.continuousdata import ContinuousData
+from eitprocessing.datahandling.eitdata import EITData
 from eitprocessing.filters import TimeDomainFilter
 
 MINUTE = 60
 NOISE_FREQUENCY_LIMIT: float = 220 / MINUTE
-NOTCH_DISTANCE: float = 10 / MINUTE
+DEFAULT_AXIS: int = 0
 
-UPPER_RESPIRATORY_RATE_LIMIT: float = 2
-UPPER_HEART_RATE_LIMIT: float = 10
+# TODO: centralize settings (these should be shared with e.g. RateDetection)
+UPPER_RESPIRATORY_RATE_LIMIT: float = 85 / MINUTE
+UPPER_HEART_RATE_LIMIT: float = 210 / MINUTE
+
+T = TypeVar("T", bound=np.ndarray | ContinuousData | EITData)
 
 
-@dataclass
+class MISSING:
+    """Sentinel value to indicate that an argument was not provided."""
+
+
+@dataclass(frozen=True, kw_only=True)
 class MDNFilter(TimeDomainFilter):
     """Multiple Digital Notch filter.
 
@@ -22,89 +34,182 @@ class MDNFilter(TimeDomainFilter):
     ± the notch distance. This is repeated for every harmonic of the heart rate below the noise
     frequency limit. Lastly, a low pass filter removes noise above the noise frequency limit.
 
-    By default, the notch distance is set to 0.166... Hz (10 BPM), and the noise frequency limit is
-    set to 3.666... Hz (220 BPM).
+    By default, the notch distance is set to 0.1̅6 Hz (10 BPM), and the noise frequency limit is
+    set to 3.̅6 Hz (220 BPM).
 
-    NB: the respiratory and heart rate should be in provided Hz, not BPM.
+    Warning:
+        The respiratory and heart rate should be in provided Hz, not BPM. We recommend defining `MINUTE = 60` and using,
+        e.g., `heart_rate=80 / MINUTE` to manually set the heart rate to 80 BPM.
 
     Args:
-      sample_frequency: the sample frequency of the data in Hz
       respiratory_rate: the respiratory rate of the subject in Hz
       heart_rate: the heart rate of the subject in Hz
       noise_frequency_limit: the highest frequency to filter in Hz
       notch_distance: the half width of the band stop filters frequency range
-
     """
 
-    sample_frequency: float
     respiratory_rate: float
     heart_rate: float
-    noise_frequency_limit: float = NOISE_FREQUENCY_LIMIT
-    notch_distance: float = NOTCH_DISTANCE
+    noise_frequency_limit: float = 220 / MINUTE
+    notch_distance: float = 10 / MINUTE
+    order: int = 10
 
     def __post_init__(self):
         if self.respiratory_rate > UPPER_RESPIRATORY_RATE_LIMIT:
             msg = (
                 f"The provided respiratory rate ({self.respiratory_rate:.1f}) "
-                f"is higher than {UPPER_RESPIRATORY_RATE_LIMIT} Hz (120 BPM). "
-                "Make sure this is correct, and to use the correct unit."
+                f"is higher than {UPPER_RESPIRATORY_RATE_LIMIT} Hz "
+                f"({UPPER_RESPIRATORY_RATE_LIMIT * MINUTE} BPM). "
+                "Make sure to use the correct unit (Hz, not BPM)."
             )
-            warnings.warn(msg)
[…]
