+++ tests/test_mdn_filter.py
@@ -1,35 +1,252 @@
+import itertools
+from collections.abc import Callable
+from functools import partial
+
 import numpy as np
-import scipy as sp
+import pytest
+from scipy import signal
 
-from eitprocessing.filters.mdn import MDNFilter
+from eitprocessing.datahandling.sequence import Sequence
+from eitprocessing.filters.mdn import UPPER_HEART_RATE_LIMIT, UPPER_RESPIRATORY_RATE_LIMIT, MDNFilter
 
-SAMPLE_FREQUENCY = 20
 MINUTE = 60
 
 
-def test_mdn_working():
-    n_samples = 10000
-    heart_rate = 70 / MINUTE
-    respiratory_rate = 12 / MINUTE
+@pytest.fixture
+def periodogram() -> Callable:
+    return partial(
+        signal.periodogram,
+        detrend="constant",
+        return_onesided=True,
+        scaling="density",
+    )
 
-    signal = np.random.normal(2.0, 1.0, n_samples)
 
-    mdn_filter = MDNFilter(sample_frequency=SAMPLE_FREQUENCY, heart_rate=heart_rate, respiratory_rate=respiratory_rate)
-    filtered_signal = mdn_filter.apply_filter(signal)
+@pytest.fixture
+def signal_factory() -> Callable[..., np.ndarray]:
+    def factory(
+        *,
+        high_power_frequencies: tuple[float, ...],
+        low_power_frequencies: tuple[float, ...],
+        duration: float = 60.0,
+        sample_frequency: float = 20.0,
+        low_power_amplitude: float = 0.2,
+        noise_amplitude: float = 0.01,
+        high_frequency_scale_factor: float = 1.0,
+        low_frequency_scale_factor: float = 1.0,
+        captures: dict | None = None,
+    ) -> np.ndarray:
+        nframes = int(duration * sample_frequency)
+        time = np.arange(nframes) / sample_frequency
+        rng = np.random.default_rng()
 
-    f, Pxx = sp.signal.welch(signal, fs=SAMPLE_FREQUENCY)
-    filtered_f, filtered_Pxx = sp.signal.welch(filtered_signal, fs=SAMPLE_FREQUENCY)
+        # Generate base signals
+        high_power_signal = np.zeros((nframes,))
+        low_power_signal = np.zeros((nframes,))
 
-    current_frequency_center = heart_rate
-    while current_frequency_center < mdn_filter.noise_frequency_limit:
-        notch_distance = mdn_filter.notch_distance
-        frequency_band = (f > current_frequency_center - notch_distance) & (
-            f < current_frequency_center + notch_distance
+        for freq in high_power_frequencies:
+            inst_freq = np.linspace(freq, freq * high_frequency_scale_factor, len(time))
+            phase = 2 * np.pi * np.cumsum(inst_freq) / sample_frequency
+            high_power_signal += rng.normal(loc=1, scale=0.1) * signal.sawtooth(phase, width=0.5)
+        for freq in low_power_frequencies:
+            inst_freq = np.linspace(freq, freq * low_frequency_scale_factor, len(time))
+            phase = 2 * np.pi * np.cumsum(inst_freq) / sample_frequency
+            low_power_signal += rng.normal(loc=1, scale=0.1) * signal.sawtooth(phase, width=0.5)
+
+        noise = rng.normal(loc=0, scale=noise_amplitude, size=(nframes,))
+
+        values = high_power_signal + low_power_amplitude * low_power_signal + noise
+
+        if captures is not None:
+            captures["high_power_signal"] = high_power_signal
[…]
