+++ docs/api/filters.md
@@ -9,3 +9,5 @@
 ::: eitprocessing.filters.butterworth_filters.BandPassFilter
 
 ::: eitprocessing.filters.butterworth_filters.ButterworthFilter
+
+::: eitprocessing.filters.mdn.MDNFilter
