+++ docs/api/filters/butterworth.md
@@ -9,5 +9,3 @@
 ::: eitprocessing.filters.butterworth_filters.BandPassFilter
 
 ::: eitprocessing.filters.butterworth_filters.ButterworthFilter
-
-::: eitprocessing.filters.mdn.MDNFilter
