+++ mkdocs.yml
@@ -65,7 +65,9 @@ nav:
         - Pixel Breath: api/features/pixel_breath.md
         - Rate Detection: api/features/rate_detection.md
         - Moving Average: api/features/moving_average.md
-      - api/filters.md
+      - Filters:
+        - Butterworth filters: api/filters/butterworth.md
+        - MDN filter: api/filters/mdn.md
       - Regions of Interest:
         - Regions of Interest: api/roi/roi.md
         - FilterROIBySize: api/roi/filterbysize.md
