### 2025-08-12 206b489
Change PixelMaskCollection to allow empty collection
