### 2025-08-12 69e76b0
Update tests for empty collection
