+++ tests/test_pixelmask_collection.py
@@ -320,17 +320,25 @@ def test_apply_with_extra_kwargs_on_array_raises(labelled_boolean_mask: Callable
         _ = collection.apply(array, sample_frequency="test")
 
 
-def test_empty_collection_raises():
-    with pytest.raises(ValueError, match="A PixelMaskCollection should contain at least one mask."):
-        _ = PixelMaskCollection([])
-
-    with pytest.raises(ValueError, match="A PixelMaskCollection should contain at least one mask."):
-        _ = PixelMaskCollection({})
-
+def test_empty_collection_behavior():
+    # Still expect TypeError if no 'masks' argument is provided at all
     with pytest.raises(
         TypeError, match=r"PixelMaskCollection.__init__\(\) missing 1 required positional argument: 'masks'"
     ):
         _ = PixelMaskCollection()
+
+    # Test both empty list and empty dict forms
+    for empty_input in ([], {}):
+        collection = PixelMaskCollection(empty_input)
+        assert len(collection.masks) == 0
+
+        # combine() should raise ValueError
+        with pytest.raises(ValueError, match="Cannot combine masks: the PixelMaskCollection is empty."):
+            collection.combine()
+
+        # apply() should raise ValueError
+        with pytest.raises(ValueError, match="Cannot apply masks: the PixelMaskCollection is empty."):
+            collection.apply(np.zeros((2, 2)))  # apply to dummy data
 
 
 def test_add_labelled_mask(labelled_boolean_mask: Callable):
