### 2025-08-12 3d3344f
Update init validation function
