### 2025-08-12 b8c362f
Update to work with empty mask collection
