### 2025-08-12 b84bc06
Update MDNFilter with make_capture
