### 2025-08-12 f050c8a
Add FilterResultsPlotting class
