+++ eitprocessing/plotting/filter.py
@@ -1,0 +1,226 @@
+from typing import TypeVar
+
+import numpy as np
+from attr import dataclass
+from matplotlib import axes, figure
+from matplotlib import pyplot as plt
+from scipy import signal
+
+from eitprocessing.datahandling.continuousdata import ContinuousData
+from eitprocessing.datahandling.eitdata import EITData
+
+T = TypeVar("T", bound=np.ndarray | ContinuousData | EITData)
+
+MINUTE = 60
+MISSING = object()
+
+
+@dataclass(frozen=True)
+class FilterPlotting:
+    """Utility class for plotting the effects of freqency filtering."""
+
+    @classmethod
+    def plot_results(
+        cls,
+        *,
+        unfiltered_data: T,
+        filtered_data: T,
+        ax: axes.Axes | None = None,
+        sample_frequency: float | object = MISSING,
+        high_pass_frequency: float | None = None,
+        low_pass_frequency: float | None = None,
+        frequency_bands: list[tuple[float, float]] | None = None,
+        xlim_to_max_filter_freq: float | None = 2,
+        **kwargs,
+    ) -> axes.Axes:
+        """Plot frequency filtering results.
+
+        This function is designed to work with the captures from a ButterworthFilter or MDNFilter.
+
+        The plot shows the power spectra for the unfiltered and filtered data. If a high pass frequency or low pass
+        frequency is provided, vertical lines are drawn at these frequencies. If frequency bands are provided, shaded
+        regions are drawn for these bands.
+
+        The provided data can be either numpy arrays, ContinuousData, or EITData, and must be the same for the
+        unfiltered and filtered data. If numpy arrays are used, the sample frequency must be provided. If ContinuousData
+        or EITData are used, the sample frequency is taken from the data itself.
+
+        If an Axes instance is provided, the plot is drawn on that Axes. If not, a new figure and axes are created.
+
+        Axes scaling:
+            The x-axis and y-axis can be scaled linearly or logarithmically using the `xscale` and `yscale` arguments.
+            If not provided, `xscale` is set to "linear" and `yscale` is set to "log". If the x-scale is logarithmic,
+            the default limits are used.
+
+            If the x-scale is linear, the upper limit is determined by the highest filter frequency or the sample
+            frequency. If xlim_to_max_filter_freq is None, the x-axis ranges from 0 to half the sample frequency. If a
+            value is provided, the x-axis limit is set to the maximum frequency of all filters multiplied by this value.
+            This enables focussing on the filtered frequencies. Custom limits can be set using the returned axes object.
+
+            If the y-scale is linear, the default limits are used. If the y-scale is logarithmic, the axes are scaled
+            such that the upper 95% of the data falls within the middle 80% of the axes. This ensures that very small
+            outliers don't disproportionally scale the y-axis.
+
+        Warning:
+            Although this function plots with the x-axis in beats per minute (BPM), the frequencies provided to it
+            should be in Hz.
+
+        Example:
+        ```
+        >>> mdn = MDNFilter(...)
+        >>> captures = {}
+        >>> filtered_data = mdn.apply(impedance_data, captures=captures)
+        >>> ax = FilterResultsPlotting().plot(**captures)
+        ```
+
+        Args:
+            unfiltered_data: The original data before filtering. Can be a numpy array, ContinuousData, or EITData.
+            filtered_data: The data after filtering. Must be the same type as unfiltered_data.
[…]
