### 2025-08-12 c611fa3
Add captures for plotting to ButterworthFilter
