### 2025-08-12 050e5b8
Add FilterPlotting to documentation, including a notebook
