+++ docs/api/plotting/filter.md
@@ -1,0 +1,1 @@
+::: eitprocessing.plotting.filter.FilterPlotting
