### 2025-08-12 b579f3c
Update RateDetection docstring
