### 2025-08-12 373810a
Update PixelMaskCollection to allow no masks argument
