### 2025-08-12 965b162
Fix testing values with all-nan array
