### 2025-08-12 02546db
Create emtpy PixelMaskCollection without argument
