### 2025-08-12 2fa9314
Re-add >>> console indicators
