### 2025-08-12 163d70e
Re-add code suggestion to save a figure
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
