+++ .github/workflows/test_build_documentation.yml
@@ -32,5 +32,5 @@ jobs:
           ln -s ../../notebooks docs/examples
           mkdocs build
       - name: Deploy documentation
-        if: github.ref == 'refs/heads/main'
-        run: mkdocs gh-deploy
+        if: ${{ github.event.action == 'push' && github.ref == 'refs/heads/main' }}
+        run: mike deploy --push --update-aliases $(git describe --tags --abbrev=0 | sed -E 's/^([vV]?[0-9]+\.[0-9]+).*/\1/') latest
