### 2025-08-12 c0ca199
Fix check for push event
