### 2025-08-13 b5489d8
Remove unnecessary suppress_value_range_error
