### 2025-08-13 e2142ce
Commit suggestion from review comments
Co-authored-by: Peter Somhorst <127968566+psomhorst@users.noreply.github.com>
