### 2025-08-13 8cce970
Add suggestion from review comments
