### 2025-08-13 1c1cd82
Change long error message to error message and note
