### 2025-08-13 eab1993
Update tests to remove unneccessary code
