### 2025-08-13 5159c0a
Remove unicode bar indicating repeating decimals
