### 2025-08-13 1e288f0
Add errors for negative rates
