### 2025-08-13 8907e55
Adapt the test so that equal rates raise an error
