### 2025-08-13 4e0ea32
Add warning about intended use of filter
