### 2025-08-13 ddffe25
Add subtraction to PixelMask class
