### 2025-08-13 7ca37c5
Added tests for 100% coverage
