### 2025-08-18 287a1ad
Add support for PixelMask plotting in PixelMap plotting
Update update method to properly update plot_config
