### 2025-08-18 23cb2bb
Add option for simulated data to Vendor
