### 2025-08-18 9e33fb0
Add function to get summed impedance to EITData
