+++ eitprocessing/datahandling/eitdata.py
@@ -4,12 +4,13 @@ import warnings
 from dataclasses import dataclass, field
 from enum import auto
 from pathlib import Path
-from typing import TYPE_CHECKING, TypeVar
+from typing import TYPE_CHECKING, Any, TypeVar
 
 import numpy as np
 from strenum import LowercaseStrEnum
 
 from eitprocessing.datahandling import DataContainer
+from eitprocessing.datahandling.continuousdata import ContinuousData
 from eitprocessing.datahandling.mixins.slicing import SelectByTime
 
 if TYPE_CHECKING:
@@ -143,6 +144,28 @@ class EITData(DataContainer, SelectByTime):
 
     def __len__(self):
         return self.pixel_impedance.shape[0]
+
+    def get_summed_impedance(self, *, return_label: str | None = None, **return_kwargs) -> ContinuousData:
+        """Return a ContinuousData-object with the same time axis and summed pixel values over time.
+
+        Args:
+            return_label: The label of the returned object; defaults to 'summed <label>' where '<label>' is the label of
+            the current object.
+            **return_kwargs: Keyword arguments for the creation of the returned object.
+        """
+        summed_impedance = np.nansum(self.pixel_impedance, axis=(1, 2))
+
+        if return_label is None:
+            return_label = f"summed {self.label}"
+
+        return_kwargs_: dict[str, Any] = {
+            "name": return_label,
+            "unit": "AU",
+            "category": "impedance",
+            "sample_frequency": self.sample_frequency,
+        } | return_kwargs
+
+        return ContinuousData(label=return_label, time=np.copy(self.time), values=summed_impedance, **return_kwargs_)
 
     def calculate_global_impedance(self) -> np.ndarray:
         """Return the global impedance, i.e. the sum of all included pixels at each frame."""
