### 2025-08-18 3248b61
Rename absolute argument to use_magnitude and add fraction_of_max argument
