### 2025-08-18 27905b3
Add captures to PixelMap.apply
