### 2025-08-18 5aa5e0c
Add IntegerMap including PlotConfig
