### 2025-08-18 dfc27bf
Add to_..._array convenience methods to PixelMap
