### 2025-08-18 7291afd
Add default PixelMapPlotConfig to registry
