### 2025-08-18 5734478
Update deprecated function name
