### 2025-08-18 cac298c
Add PixelMask.values alias to .mask
