### 2025-08-18 d43f8ff
Add PixelMask.shape property
