### 2025-08-18 ca952fb
Add __eq__ method to check equivalence of PixelMasks
