### 2025-08-18 1fc168b
Replace fixed default masks with generated masks, based on shape
A generator was necessary anyway for circular import reasons. While
writing, I remembered I wanted to write a function for non-standard
shapes anyway. This commit kills two birds with one stone.
