+++ eitprocessing/roi/__init__.py
@@ -23,6 +23,7 @@ vice versa.
 
 from __future__ import annotations
 
+import re
 import sys
 import warnings
 from dataclasses import InitVar, dataclass, field, replace
@@ -294,18 +295,118 @@ class PixelMask:
         return self.mask.shape == other.mask.shape and np.array_equal(self.mask, other.mask, equal_nan=True)
 
 
-LAYER_1_MASK = PixelMask(np.concat([np.ones((8, 32)), np.full((24, 32), np.nan)], axis=0))
-LAYER_2_MASK = PixelMask(np.concat([np.full((8, 32), np.nan), np.ones((8, 32)), np.full((16, 32), np.nan)], axis=0))
-LAYER_3_MASK = PixelMask(np.concat([np.full((16, 32), np.nan), np.ones((8, 32)), np.full((8, 32), np.nan)], axis=0))
-LAYER_4_MASK = PixelMask(np.concat([np.full((24, 32), np.nan), np.ones((8, 32))], axis=0))
+def get_geometric_mask(mask: str, shape: tuple[int, int] = (32, 32)) -> PixelMask:
+    """Get a geometric mask by name.
 
-VENTRAL_MASK = LAYER_1_MASK + LAYER_2_MASK
-DORSAL_MASK = LAYER_3_MASK + LAYER_4_MASK
+    Masks can be generated for appropriates shapes, provided by the shape` argument, a tuple of two integers
+    representing the number of rows and columns in the EIT image. The shape has to be divisible in the required number
+    of
 
-ANATOMICAL_RIGHT_MASK = PixelMask(np.concat([np.ones((32, 16)), np.full((32, 16), np.nan)], axis=1))
-ANATOMICAL_LEFT_MASK = PixelMask(np.concat([np.full((32, 16), np.nan), np.ones((32, 16))], axis=1))
+    The function accepts both full names (e.g., "layer 1") and abbreviations (e.g., "L1").
 
-QUADRANT_1_MASK = VENTRAL_MASK * ANATOMICAL_RIGHT_MASK
-QUADRANT_2_MASK = VENTRAL_MASK * ANATOMICAL_LEFT_MASK
-QUADRANT_3_MASK = DORSAL_MASK * ANATOMICAL_RIGHT_MASK
-QUADRANT_4_MASK = DORSAL_MASK * ANATOMICAL_LEFT_MASK
+    The following masks are available:
+    - "ventral" or "V": the first half rows of the EIT image.
+    - "dorsal" or "D": the last half rows of the EIT image.
+    - "anatomical right" or "R": the first half columns of the EIT image.
+    - "anatomical left" or "L": the last half columns of the EIT image.
+    - "layer 1" or "L1": the first quarter rows of the EIT image.
+    - "layer 2" or "L2": the second set of quarter rows of the EIT image.
+    - "layer 3" or "L3": the third set of quarter rows of the EIT image.
+    - "layer 4" or "L4": the last quarter rows of the EIT image.
+    - "quadrant 1" or "Q1": the top right quadrant of the EIT image.
+    - "quadrant 2" or "Q2": the top left quadrant of the EIT image.
+    - "quadrant 3" or "Q3": the bottom right quadrant of the EIT image.
+    - "quadrant 4" or "Q4": the bottom left quadrant of the EIT image.
+
+    Args:
+        mask: Name of the geometric mask to retrieve (case-sensitive).
+        shape:
+            Shape of the EIT image, a tuple of two integers representing the number of rows and columns. Defaults to
+            (32, 32).
+
+    Returns:
+        PixelMask: The requested geometric mask.
+
+    Raises:
+        ValueError: If an unknown mask name is provided.
+        ValueError: If the shape is not compatible with the requested mask.
+    """
+
[…]
