### 2025-08-18 f24510b
Add contour and surface plots to PixelMap
