+++ eitprocessing/plotting/pixelmap.py
@@ -33,15 +33,18 @@ from dataclasses import dataclass, field
 import matplotlib as mpl
 import numpy as np
 from frozendict import frozendict
-from matplotlib import colorbar
+from matplotlib import axes, colorbar
 from matplotlib import pyplot as plt
 from matplotlib.axes import Axes
 from matplotlib.colors import BoundaryNorm, CenteredNorm, Colormap, LinearSegmentedColormap, ListedColormap, Normalize
 from matplotlib.image import AxesImage
 from matplotlib.ticker import PercentFormatter
+from mpl_toolkits import mplot3d
+from mpl_toolkits.mplot3d.art3d import Poly3DCollection
+from scipy import ndimage
 
 from eitprocessing.config import Config
-from eitprocessing.datahandling.pixelmap import PixelMap
+from eitprocessing.datahandling.pixelmap import IntegerMap, PixelMap
 from eitprocessing.plotting.helpers import AbsolutePercentFormatter, AbsoluteScalarFormatter
 from eitprocessing.roi import PixelMask
 
@@ -185,6 +188,161 @@ class PixelMapPlotting:
             colorbar_kwargs.setdefault("extend", extend)
 
         return plt.colorbar(cm, ax=ax, **colorbar_kwargs or {})
+
+    def contour(self, **kwargs) -> mpl.contour.ContourSet:
+        """Create a contour plot of the pixel map or mask.
+
+        This method uses `matplotlib.pyplot.contour` to create a contour plot of the pixel map values.
+
+        Args:
+            **kwargs: Additional keyword arguments passed to `matplotlib.pyplot.contour`.
+
+        Returns:
+            QuadContourSet: The contour set created by the contour function.
+        """
+        ax = kwargs.pop("ax", plt.gca())
+        ax.set(aspect=kwargs.pop("aspect", "equal"))
+
+        origin = kwargs.pop("origin", "upper")
+
+        if origin not in ("upper", "lower"):
+            msg = f"Invalid origin '{origin}'. Expected 'upper' or 'lower'."
+            raise ValueError(msg)
+
+        if origin == "upper":
+            ax.invert_yaxis()
+
+        y_grid, x_grid = np.indices(self.pixel_map.shape)
+        return ax.contour(x_grid, y_grid, self.pixel_map.values, **kwargs)
+
+    def surface(self, **kwargs) -> Poly3DCollection:
+        """Create a 3D surface plot of the pixel map or mask.
+
+        This method uses `matplotlib.pyplot.axes3d.plot_surface` to create a 3D surface plot of the pixel map values.
+
+        Args:
+            **kwargs: Additional keyword arguments passed to `matplotlib.pyplot.axes3d.plot_surface`.
+
+        Returns:
+            Axes3D: The 3D axes object containing the surface plot.
+        """
+        ax = kwargs.pop("ax", None)
+        if ax is not None:
+            if not isinstance(ax, mplot3d.Axes3D):
+                msg = "Expected ax to be a 3D Axes object."
+                raise TypeError(msg)
+        else:
+            fig = plt.figure()
+            ax = fig.add_subplot(111, projection="3d")
+
+        ax.view_init(elev=45, azim=10)
+        y_grid, x_grid = np.indices(self.pixel_map.shape)
+        return ax.plot_surface(
+            x_grid,
+            y_grid,
+            self.pixel_map.values,
+            cmap=kwargs.pop("cmap", self.config.cmap),
+            vmin=kwargs.pop("vmin", 0),
+            **kwargs,
+        )
+
+    def add_region_markers(self, ax: axes.Axes, label_map: dict | None = None, **kwargs) -> list[mpl.text.Text]:
+        """Add markers to an integer map plot.
+
+        This methods adds text labels at the center of all pixels with the same label, assuming continuous regions where
[…]
