### 2025-08-18 bdb9b1c
Update tests for PixelMasks
