+++ tests/test_pixelmask.py
@@ -5,21 +5,7 @@ import pytest
 
 from eitprocessing.datahandling.pixelmap import PixelMap
 from eitprocessing.datahandling.sequence import Sequence
-from eitprocessing.roi import (
-    ANATOMICAL_LEFT_MASK,
-    ANATOMICAL_RIGHT_MASK,
-    DORSAL_MASK,
-    LAYER_1_MASK,
-    LAYER_2_MASK,
-    LAYER_3_MASK,
-    LAYER_4_MASK,
-    QUADRANT_1_MASK,
-    QUADRANT_2_MASK,
-    QUADRANT_3_MASK,
-    QUADRANT_4_MASK,
-    VENTRAL_MASK,
-    PixelMask,
-)
+from eitprocessing.roi import PixelMask, get_geometric_mask
 
 
 def test_pixelmask_init_with_boolean_array():
@@ -234,22 +220,47 @@ def test_pixelmask_subtract():
     assert np.array_equal(pm3.mask, np.array([[np.nan, np.nan, 1, np.nan], [0.1, np.nan, np.nan, 0.5]]), equal_nan=True)
 
 
-def test_predefined_layer_masks():
-    assert np.all(LAYER_1_MASK.mask[:8, :] == 1.0)
-    assert np.all(np.isnan(LAYER_1_MASK.mask[8:, :]))
+@pytest.mark.parametrize("shape", [(32, 1), (64, 1), (16, 1), (100, 1), (4, 1)])
+def test_predefined_layer_masks(shape: tuple[int, int]):
+    quarter_height = shape[0] // 4
+    first = slice(None, quarter_height)
+    second = slice(quarter_height, 2 * quarter_height)
+    third = slice(2 * quarter_height, 3 * quarter_height)
+    fourth = slice(3 * quarter_height, None)
+    layer1_mask = get_geometric_mask("layer 1", shape)
 
-    assert np.all(LAYER_2_MASK.mask[8:16, :] == 1.0)
-    assert np.all(np.isnan(LAYER_2_MASK.mask[:8, :]))
-    assert np.all(np.isnan(LAYER_2_MASK.mask[16, :]))
+    assert layer1_mask.label == "layer 1"
+    assert layer1_mask == get_geometric_mask("L1", shape)
+    assert get_geometric_mask("L1").label == "layer 1"
+    assert layer1_mask.shape == shape
 
-    assert np.all(LAYER_3_MASK.mask[16:24, :] == 1.0)
-    assert np.all(np.isnan(LAYER_3_MASK.mask[:16, :]))
-    assert np.all(np.isnan(LAYER_3_MASK.mask[24, :]))
+    assert np.all(layer1_mask.mask[first, :] == 1.0)
+    assert np.all(np.isnan(layer1_mask.mask[second, :]))
+    assert np.all(np.isnan(layer1_mask.mask[third, :]))
+    assert np.all(np.isnan(layer1_mask.mask[fourth, :]))
 
-    assert np.all(LAYER_4_MASK.mask[24:, :] == 1.0)
-    assert np.all(np.isnan(LAYER_4_MASK.mask[:24, :]))
+    layer2_mask = get_geometric_mask("layer 2", shape)
+    assert layer2_mask.label == "layer 2"
+    assert np.all(np.isnan(layer2_mask.mask[first, :]))
+    assert np.all(layer2_mask.mask[second, :] == 1.0)
+    assert np.all(np.isnan(layer2_mask.mask[third, :]))
+    assert np.all(np.isnan(layer2_mask.mask[fourth, :]))
 
[…]
