### 2025-08-18 513f28b
Update tests for PixelMap
