### 2025-08-18 69f5c2b
Add WatershedLungspace class and accompanying tests
