+++ eitprocessing/roi/watershed.py
@@ -1,0 +1,170 @@
+from dataclasses import dataclass
+
+import numpy as np
+from scipy import ndimage as ndi
+from skimage import feature as ski_feature
+from skimage import segmentation as ski_segmentation
+
+from eitprocessing.datahandling.continuousdata import ContinuousData
+from eitprocessing.datahandling.eitdata import EITData
+from eitprocessing.datahandling.pixelmap import IntegerMap, PixelMap, TIVMap
+from eitprocessing.features.pixel_breath import PixelBreath
+from eitprocessing.parameters.tidal_impedance_variation import TIV
+from eitprocessing.roi import PixelMask
+from eitprocessing.utils import make_capture
+
+
+@dataclass(kw_only=True, frozen=True)
+class WatershedLungspace:
+    """Create a pixel mask based on the watershed method.
+
+    This method was designed to improve functional lung space detection in the presence of pendelluft. Functional lung
+    space defined as pixels with a tidal impedance variation (TIV) at or above a percentage of the maximum TIV can
+    result in underdetection of pixels that have reduced TIV due to the pendelluft phenomenon. An alternate approach
+    using the pixel amplitude instead of the TIV results in overinclusion of pixels.
+
+    In this method, the (inverse) watershed method is applied to the pixel amplitude. This results in distinct regions
+    with high values for more central pixels, and low values for edge values. Regions where the highest value falls
+    within the TIV-based functional lung space definition are included in the mask. Other regions are exluded. Pixels
+    that fall outside the amplitude based functional lung space definiton are excluded.
+
+    Example usage:
+    ```python
+    >>> mask = WatershedLungspace(threshold_fraction=0.15).apply(eit_data)
+    >>> masked_eit_data = mask.apply(eit_data)
+    ```
+
+    Args:
+        threshold_fraction (float):
+            The fraction of the maximum TIV for the initial functional lung space definition that is used in the
+            algorithm. Defaults to 0.15 (15%).
+    """
+
+    threshold_fraction: float = 0.15
+
+    def __post_init__(self):
+        if not isinstance(self.threshold_fraction, float):
+            msg = "Threshold must be a float."
+            raise TypeError(msg)
+        if not (0 < self.threshold_fraction < 1):
+            msg = "Threshold must be between 0 and 1."
+            raise ValueError(msg)
+
+    def apply(
+        self, eit_data: EITData, *, timing_data: ContinuousData | None = None, captures: dict | None = None
+    ) -> PixelMask:
+        """Apply the watershed method to the EIT data.
+
+        `BreathDetection` is used to find breaths in timing data. By default, the timing data is the summed pixel
+        impedance. Alternative timing data, e.g., pressure data, can be provided.
+
+        `TIV` is used to compute the mean tidal impedance variation (TIV) and mean pixel amplitude over the breaths. The
+        absolute threshold is defined as 15% of the maximum TIV over all pixels. A functional TIV mask and functional
+        amplitude mask are defined as all pixels with a TIV/amplitude of at least the absolute threshold.
+
+        Local peaks in the amplitude map are found using `skimage.feature.peak_local_max`. These peaks are used to find
+        the (inverse) watershed regions in the amplitude map. Regions whose peaks fall inside the functional TIV mask
+        are included, the others are exluded. The final watershed mask is the intersection of the functional amplitude
+        mask and the remaining included watershed regions.
+
+        Watershed regions:
+            The concept of watershed regions come from geography, where it refers to the area of land that drains into a
+            single river or body of water. In the context of image processing, it refers to the region of an image that
[…]
