### 2025-08-18 c5bca1f
Limit version of pytest
pytest changed fixture wrapping in 8.4.0, which changes how fixutures
can be accessed outside of tests.
