+++ pyproject.toml
@@ -51,14 +51,14 @@ dependencies = [
 [project.optional-dependencies]
 # development dependency groups
 dev = [
-    "pytest >= 7.4.0",
+    "pytest >= 7.4.0,< 8.4.0",
     "bump-my-version==0.28.1",
     "coverage",
     "pytest-cov",
     "coveralls",
     "ruff==0.11.7",
 ]
-testing = ["pytest >= 7.4.0", "pytest-cov", "coveralls"]
+testing = ["pytest >= 7.4.0, < 8.4.0", "pytest-cov", "coveralls"]
 docs = [
     "mkdocs",
     "mkdocs-material",
