### 2025-08-18 e4ff11d
Set stack level and type of warnings
This makes warnings more readable in general
