### 2025-08-18 5bac9b0
Add watershed and notebook to documentation
