### 2025-08-18 a67d7fe
Update PixelMaskCollection tests to use real EITData
