### 2025-08-18 3067b7a
Add test to compare performance of uv
