### 2025-08-21 3df5ac5
Remove pip workflow and rename uv workflow
