+++ .github/actions/install_eitprocessing_uv/action.yml
@@ -1,48 +1,0 @@
-name: Install eitprocessing
-description: Install eitprocessing on the given Python version.
-inputs:
-  python-version:
-    description: "The Python version to use"
-    required: false
-    default: "3.10"
-  dependencies:
-    description: "The optional dependencies of eitprocessing to install"
-    required: false
-  extract-data:
-    description: "Whether to extract testing data"
-    required: false
-    default: "false"
-  token:
-    description: "GitHub TOKEN"
-    required: true
-  data-directory:
-    description: "Directory where to store eitprocessing data"
-    required: false
-    default: ${{ github.workspace }}/../eitprocessing_data/
-
-runs:
-  using: "composite"
-  steps:
-    - uses: actions/setup-python@v5
-      with:
-        python-version: ${{ inputs.python-version }}
-    - uses: docker/login-action@v3
-      if: ${{ inputs.extract-data == 'true' }}
-      with:
-        registry: ghcr.io
-        username: psomhorst
-        password: ${{ inputs.token }}
-    - uses: shrink/actions-docker-extract@v3
-      if: ${{ inputs.extract-data == 'true' }}
-      with:
-        image: ghcr.io/eit-alive/eittestdata:latest
-        destination: ${{ inputs.data-directory }}
-        path: /eitprocessing/.
-    - name: Install uv and set the python version
-      uses: astral-sh/setup-uv@v6
-      with:
-        enable-cache: true
-    - name: Install the project
-      run: uv sync ${{ inputs.dependencies }}
-      shell: bash
-
