+++ .github/workflows/testing.yml
@@ -1,4 +1,4 @@
-name: Build and Test
+name: Run tests
 
 on:
   push:
@@ -32,10 +32,10 @@ jobs:
       - uses: actions/checkout@v4
       - uses: ./.github/actions/install_eitprocessing
         with:
-          dependencies: testing
+          dependencies: --extra testing
           extract-data: true
           python-version: ${{ matrix.python-version }}
           token: ${{ secrets.GITHUB_TOKEN }}
           data-directory: ${{ env.EIT_PROCESSING_TEST_DATA }}
       - name: Run pytest
-        run: pytest -v
+        run: uv run pytest -v
