+++ .github/workflows/testing_uv.yml
@@ -1,41 +1,0 @@
-name: Build and Test (uv)
-
-on:
-  push:
-    branches:
-      - main
-      - develop
-      - ci/uv
-  pull_request:
-    types: [opened, synchronize, reopened, ready_for_review]
-    paths-ignore:
-      - "**.md"
-      - "**.rst"
-      - "**.ipynb"
-    branches:
-      - main
-      - develop
-
-jobs:
-  test_and_build:
-    if: github.event.pull_request.draft == false
-    runs-on: ${{ matrix.os }}
-    strategy:
-      fail-fast: false
-      matrix:
-        os: ["ubuntu-latest"]
-        python-version: ["3.10"]
-    name: Test and build for ${{ matrix.python-version }}, ${{ matrix.os }}
-    env:
-      EIT_PROCESSING_TEST_DATA: ${{ github.workspace }}/../eitprocessing_data/
-    steps:
-      - uses: actions/checkout@v4
-      - uses: ./.github/actions/install_eitprocessing_uv
-        with:
-          dependencies: --extra testing
-          extract-data: true
-          python-version: ${{ matrix.python-version }}
-          token: ${{ secrets.GITHUB_TOKEN }}
-          data-directory: ${{ env.EIT_PROCESSING_TEST_DATA }}
-      - name: Run pytest
-        run: uv run pytest -v
