### 2025-08-21 78f578b
All equivalent classes should have hash functions
