### 2025-08-21 b12c1ea
Ignore circular import issues where necessary
