+++ eitprocessing/datahandling/pixelmap.py
@@ -150,7 +150,7 @@ class PixelMap:
         if plot_config is None:
             plot_config = {}
         if isinstance(plot_config, dict):
-            from eitprocessing.plotting import get_plot_config
+            from eitprocessing.plotting import get_plot_config  # noqa: PLC0415
 
             default_config = get_plot_config(self)
             plot_config = default_config.update(**plot_config)
@@ -353,7 +353,7 @@ class PixelMap:
             msg = "`comparator` must be a callable function."
             raise TypeError(msg)
 
-        from eitprocessing.roi import PixelMask
+        from eitprocessing.roi import PixelMask  # noqa: PLC0415
 
         compare_values = np.abs(self.values) if use_magnitude else self.values
 
