### 2025-08-21 a5783cc
Add match strings to caught deprecations warnings in tests
