### 2025-08-21 51ca669
Add config for marking tests as slow
