### 2025-08-21 28f77e2
Remove superfluous text in github action
