+++ .github/actions/install_eitprocessing/action.yml
@@ -38,7 +38,7 @@ runs:
         image: ghcr.io/eit-alive/eittestdata:latest
         destination: ${{ inputs.data-directory }}
         path: /eitprocessing/.
-    - name: Install uv and set the python version
+    - name: Install uv
       uses: astral-sh/setup-uv@v6
       with:
         enable-cache: true
