### 2025-08-21 7c6ba19
Lint using uv as installer
