### 2025-08-21 d50c883
Release using uv as installer
