### 2025-08-22 313a742
Add guard against leaving temporary workflow runs
