### 2025-08-22 4e0ff70
Improve IntegerMap with dtype class variable and lossless conversion checks
