### 2025-08-22 dff9811
Add test value conversion at init
