### 2025-08-22 207321f
Update IntegerMap plot config
