### 2025-08-22 48b3dc9
Update to_boolean_array to optionally keep zero values as True
