### 2025-08-22 d6e6cb4
Update pixelmap tests
