### 2025-08-22 17a8530
Add check for no markers inside TIV mask
