### 2025-08-22 672e47d
Raise error if no breaths are found in breath detection
