### 2025-08-22 7e82552
Add test for watershed with no amplitude
