### 2025-08-22 95af7b4
Add test for captures
