### 2025-08-25 c85340e
Add TIVLungspace class
