### 2025-08-25 f2fb3a6
Add tests for TIVLungspace class
