+++ eitprocessing/roi/watershed.py
@@ -8,9 +8,8 @@ from skimage import segmentation as ski_segmentation
 from eitprocessing.datahandling.continuousdata import ContinuousData
 from eitprocessing.datahandling.eitdata import EITData
 from eitprocessing.datahandling.pixelmap import IntegerMap, PixelMap, TIVMap
-from eitprocessing.features.pixel_breath import PixelBreath
-from eitprocessing.parameters.tidal_impedance_variation import TIV
 from eitprocessing.roi import PixelMask
+from eitprocessing.roi.tiv import TIVLungspace
 from eitprocessing.utils import make_capture
 
 
@@ -23,7 +22,14 @@ class WatershedLungspace:
     result in underdetection of pixels that have reduced TIV due to the pendelluft phenomenon. An alternate approach
     using the pixel amplitude instead of the TIV results in overinclusion of pixels.
 
-    In this method, the (inverse) watershed method is applied to the pixel amplitude. This results in distinct regions
+    Watershed regions:
+        The concept of watershed regions come from geography, where it refers to the area of land that drains into a
+        single river or body of water. In the context of image processing, it refers to the region of an image that
+        is associated with a particular local minimum. The region borders are defined by the 'ridges' between the
+        local minima. The inverse watershed method uses maxima and 'valleys' between them. This algorithm uses the
+        inverse method, where local maximum impedance values form the centers of watershed regions.
+
+    In this method, the inverse watershed method is applied to the pixel amplitude. This results in distinct regions
     with high values for more central pixels, and low values for edge pixels. Regions where the highest value falls
     within the TIV-based functional lung space definition are included in the mask. Other regions are excluded. Pixels
     that fall outside the amplitude based functional lung space definition are excluded.
@@ -50,29 +56,21 @@ class WatershedLungspace:
             msg = "Threshold must be between 0 and 1."
             raise ValueError(msg)
 
-    def apply(
+    def apply(  # noqa: PLR0915
         self, eit_data: EITData, *, timing_data: ContinuousData | None = None, captures: dict | None = None
     ) -> PixelMask:
         """Apply the watershed method to the EIT data.
 
-        `BreathDetection` is used to find breaths in timing data. By default, the timing data is the summed pixel
-        impedance. Alternative timing data, e.g., pressure data, can be provided.
-
-        `TIV` is used to compute the mean tidal impedance variation (TIV) and mean pixel amplitude over the breaths. The
-        absolute threshold is defined as 15% of the maximum TIV over all pixels. A functional TIV mask and functional
-        amplitude mask are defined as all pixels with a TIV/amplitude of at least the absolute threshold.
+        TIVLungspace is used to gather the mean TIV and mean amplitude maps, and their associated functional lung space
+        masks. The lung space masks are redefined here to ensure that the same breaths are used for both TIV and
+        amplitude.
 
         Local peaks in the amplitude map are found using `skimage.feature.peak_local_max`. These peaks are used to find
         the (inverse) watershed regions in the amplitude map. Regions whose peaks fall inside the functional TIV mask
         are included, the others are excluded. The final watershed mask is the intersection of the functional amplitude
         mask and the remaining included watershed regions.
 
-        Watershed regions:
-            The concept of watershed regions come from geography, where it refers to the area of land that drains into a
-            single river or body of water. In the context of image processing, it refers to the region of an image that
-            is associated with a particular local minimum. The region borders are defined by the 'ridges' between the
-            local minima. The inverse watershed method uses maxima and 'valleys' between them. This algorithm uses the
[…]
