### 2025-08-25 2b59718
Move mkdocs-jupyter dependency to docs group
