+++ pyproject.toml
@@ -45,7 +45,6 @@ dependencies = [
     "typing_extensions",
     "pyyaml",
     "frozendict>=2.4.6",
-    "mkdocs-jupyter>=0.25.1",
     "scikit-image>=0.25.2",
 ]
 
@@ -69,6 +68,7 @@ docs = [
     "mike",
     "pymdown-extensions",
     "black",
+    "mkdocs-jupyter >=0.25.1",
 ]
 notebooks = [
     "jupyter >= 1.0.0",
