### 2025-08-25 220f4f2
Add version info for docs dependencies
