+++ pyproject.toml
@@ -60,14 +60,14 @@ dev = [
 ]
 testing = ["pytest >= 7.4.0, < 8.4.0", "pytest-cov", "coveralls"]
 docs = [
-    "mkdocs",
-    "mkdocs-material",
-    "mkdocstrings",
-    "mkdocstrings-python",
-    "mkdocs-exclude",
-    "mike",
-    "pymdown-extensions",
-    "black",
+    "mkdocs >= 1.6.1",
+    "mkdocs-material >= 9.6.18",
+    "mkdocstrings >= 0.30.0",
+    "mkdocstrings-python >= 1.17.0",
+    "mkdocs-exclude >= 1.0.2",
+    "mike >= 2.1.3",
+    "pymdown-extensions >= 10.16.1",
+    "black >= 25.1.0",
     "mkdocs-jupyter >=0.25.1",
 ]
 notebooks = [
