### 2025-08-25 9d69897
Create separate AmplitudeLungspace class, and remove modes from TIVLungspace
