+++ eitprocessing/roi/amplitude.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from eitprocessing.datahandling.continuousdata import ContinuousData
 from eitprocessing.datahandling.eitdata import EITData
-from eitprocessing.datahandling.pixelmap import TIVMap
+from eitprocessing.datahandling.pixelmap import AmplitudeMap
 from eitprocessing.features.pixel_breath import PixelBreath
 from eitprocessing.parameters.tidal_impedance_variation import TIV
 from eitprocessing.roi import PixelMask
@@ -13,20 +13,20 @@ from eitprocessing.utils import make_capture
 
 
 @dataclass(kw_only=True, frozen=True)
-class TIVLungspace:
-    """Create a pixel mask by thresholding the mean TIV.
+class AmplitudeLungspace:
+    """Create a pixel mask by thresholding the mean amplitude.
 
-    This defines the functional lung space as all pixels with a tidal impedance variation (TIV) of at least the provided
-    fractional threshold of the maximum TIV.
+    This defines the functional lung space as all pixels with an amplitude of at least the provided fractional threshold
+    of the maximum ampltiude.
 
     Example usage:
     ```python
-    >>> mask = TIVLungspace(threshold=0.15).apply(eit_data)
+    >>> mask = AmplitudeLungspace(threshold=0.15).apply(eit_data)
     >>> masked_eit_data = mask.apply(eit_data)
     ```
 
     Args:
-        threshold (float): The fraction of the maximum TIV that is used as threshold. Defaults to 0.15 (15%).
+        threshold (float): The fraction of the maximum amplitude that is used as threshold. Defaults to 0.15 (15%).
     """
 
     threshold: float = 0.15
@@ -42,13 +42,13 @@ class TIVLungspace:
     def apply(
         self, eit_data: EITData, *, timing_data: ContinuousData | None = None, captures: dict | None = None
     ) -> PixelMask:
-        """Apply the TIV thresholding to the EIT data.
+        """Apply the amplitude thresholding to the EIT data.
 
         `BreathDetection` is used to find breaths in timing data. By default, the timing data is the summed pixel
         impedance. Alternative timing data, e.g., pressure data, can be provided.
 
-        Then, `TIV` is used to compute the TIV for each breath. The mean TIV over all breaths is computed,
-        and pixels with a mean TIV above the threshold are included in the mask.
+        Then, `TIV` is used to compute the amplitude for each breath. The mean amplitude over all breaths is computed,
+        and pixels with a mean amplitude above the threshold are included in the mask.
 
         Args:
             eit_data (EITData): The EIT data to process.
@@ -64,28 +64,28 @@ class TIVLungspace:
         if timing_data is None:
             timing_data = eit_data.get_summed_impedance()
 
-        tiv_per_breath = TIV(
-            pixel_breath=PixelBreath(phase_correction_mode="negative amplitude")
+        amplitude_per_breath = TIV(
+            pixel_breath=PixelBreath(phase_correction_mode="phase shift")
         ).compute_pixel_parameter(
             eit_data=eit_data,
             continuous_data=timing_data,
-            tiv_timing="continuous",
+            tiv_timing="pixel",
             sequence=None,
         )
-        if not len(tiv_per_breath):
-            msg = "No breaths were detected. Cannot compute TIV."
+        if not len(amplitude_per_breath):
+            msg = "No breaths were detected. Cannot compute amplitude."
             raise ValueError(msg)
 
-        if np.all(np.isnan(tiv_per_breath.values)):
-            msg = "No non-nan TIV values were found."
+        if np.all(np.isnan(amplitude_per_breath.values)):
+            msg = "No non-nan amplitude values were found."
[…]
