### 2025-08-25 a9042fe
Update documentation for AmplitudeLungspace
