+++ docs/api/roi/tivlungspace.md
@@ -1,1 +1,3 @@
 ::: eitprocessing.roi.tiv.TIVLungspace
+
+::: eitprocessing.roi.amplitude.AmplitudeLungspace
