+++ mkdocs.yml
@@ -74,8 +74,8 @@ nav:
         - Regions of Interest: api/roi/roi.md
         - FilterROIBySize: api/roi/filterbysize.md
         - PixelMaskCollection: api/roi/pixelmaskcollection.md
-        - TIVLungspace: api/roi/tivlunspace.md
-        - Watershed Lungspace: api/roi/watershed.md
+        - TIV and amplitude lung space: api/roi/tivlungspace.md
+        - Watershed lung space: api/roi/watershed.md
       - api/parameters.md
       - api/categories.md
       - api/mixins.md
