### 2025-08-25 fedebb7
Add AmplitudeMap, AmplitudeMapPlotConfig
