### 2025-08-25 dd072d5
Re-add warning for AmplitudeLungspace
