### 2025-08-25 1c56417
Update tests for AmplitudeLungspace
