+++ tests/test_amplitude_lungspace.py
@@ -6,7 +6,7 @@ from numpy import typing as npt
 
 from eitprocessing.datahandling.eitdata import EITData, Vendor
 from eitprocessing.datahandling.sequence import Sequence
-from eitprocessing.roi.tiv import TIVLungspace
+from eitprocessing.roi.amplitude import AmplitudeLungspace
 
 
 @pytest.fixture
@@ -42,32 +42,34 @@ def create_signal():
 
 
 def test_tiv_lungspace_init():
-    _ = TIVLungspace()
-    _ = TIVLungspace(threshold=0.2)
+    _ = AmplitudeLungspace()
+    _ = AmplitudeLungspace(threshold=0.2)
 
 
 @pytest.mark.parametrize("threshold", [-0.1, 0.0, 1.0, 1.5])
 def test_tiv_lungspace_init_threshold_outside_range(threshold: float):
     with pytest.raises(ValueError, match="Threshold must be between 0 and 1."):
-        _ = TIVLungspace(threshold=threshold)
+        _ = AmplitudeLungspace(threshold=threshold)
 
 
 @pytest.mark.parametrize("threshold", ["0.2", True, None])
 def test_tiv_lungspace_init_threshold_wrong_type(threshold: object):
     with pytest.raises(TypeError, match="Threshold must be a float."):
-        _ = TIVLungspace(threshold=threshold)
+        _ = AmplitudeLungspace(threshold=threshold)
 
 
 def test_tiv_lungspace_apply(create_signal: Callable):
     amplitudes = np.array([[-1.0, -0.5, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0]])
     signal = create_signal(amplitudes, duration=60)
-    tiv_lungspace = TIVLungspace(threshold=0.15).apply(signal, captures=(captures_tiv := {}))
+    tiv_lungspace = AmplitudeLungspace(threshold=0.15).apply(signal, captures=(captures_tiv := {}))
 
     assert tiv_lungspace.values.shape == amplitudes.shape
-    assert np.allclose(captures_tiv["mean TIV"].values, [[-2.0, -1.0, 0, 0.4, 0.8, 1.2, 1.6, 2.0]])
+    assert np.allclose(
+        captures_tiv["mean amplitude"].values, [[2.0, 1.0, np.nan, 0.4, 0.8, 1.2, 1.6, 2.0]], equal_nan=True
+    )
     assert np.array_equal(
         tiv_lungspace.values,
-        np.array([[np.nan, np.nan, np.nan, 1.0, 1.0, 1.0, 1.0, 1.0]]),
+        np.array([[1.0, 1.0, np.nan, 1.0, 1.0, 1.0, 1.0, 1.0]]),
         equal_nan=True,
     )
 
@@ -76,16 +78,17 @@ def test_tiv_lungspace_apply_no_breaths(create_signal: Callable):
     amplitudes = np.zeros((1, 8))
     signal = create_signal(amplitudes, duration=60)
 
-    with pytest.raises(ValueError, match="No breaths were detected. Cannot compute TIV"):
-        _ = TIVLungspace(threshold=0.2).apply(signal)
+    with pytest.raises(ValueError, match="No breaths were detected. Cannot compute amplitude."):
+        _ = AmplitudeLungspace(threshold=0.2).apply(signal)
 
 
 def test_tiv_lungspace_apply_no_non_nan_values(create_signal: Callable):
     amplitudes = np.full((1, 8), 1)
     signal = create_signal(amplitudes, duration=15)
 
-    # Works with TIV (but not amplitude), because it needs only one detected breath
-    _ = TIVLungspace(threshold=0.2).apply(signal)
[…]
