### 2025-08-25 1b434c3
Add AmplitudeMap and IntegerMap to pixelmap docs
