### 2025-08-25 51b42ef
Only consider pixels that are not all-NaN when aggregating
