### 2025-08-25 3daeba8
Suppress zero warning when subtracting masks
