### 2025-08-25 b909af1
Add exception when trying to normalize a all-NaN PixelMap
