### 2025-08-25 650c5f2
Add explicit check for all-NaN values
