### 2025-08-25 f45fc39
Test butterworth filters with apply instead of deprecated apply_filter
