### 2025-08-25 9c33751
Catch or prevent warnings in tests
