+++ tests/test_pixelmap.py
@@ -9,6 +9,7 @@ import pytest
 from matplotlib import pyplot as plt
 from matplotlib.colors import CenteredNorm, Colormap, Normalize
 from matplotlib.ticker import PercentFormatter, ScalarFormatter
+from numpy.exceptions import ComplexWarning
 
 from eitprocessing.datahandling.pixelmap import (
     DifferenceMap,
@@ -321,8 +322,8 @@ def test_normalize_values_errors():
     with pytest.raises(ValueError, match="Normalization by NaN is not allowed"):
         _ = pm.normalize(mode="reference", reference=np.nan)
 
-    with pytest.raises(ValueError, match="Normalization by NaN is not allowed"):
-        _ = PixelMap([[np.nan]]).normalize()
+    with pytest.raises(ValueError, match="Cannot normalize a PixelMap with all NaN values."):
+        _ = PixelMap([[np.nan]], suppress_all_nan_warning=True).normalize()
 
     with pytest.warns(UserWarning, match="Normalization by a negative number"):
         _ = pm.normalize(mode="reference", reference=-1)
@@ -437,7 +438,10 @@ def test_add():
 
 def test_sub():
     pm1 = PerfusionMap([[0, 1, 2, 3]], plot_config={"absolute": True})
-    pm1_sub = pm1 - 3
+    with pytest.warns(
+        UserWarning, match="PerfusionMap initialized with negative values, but `allow_negative_values` is False."
+    ):
+        pm1_sub = pm1 - 3
     assert np.array_equal(pm1_sub.values, np.array([[-3, -2, -1, 0]]))
     assert isinstance(pm1_sub, PerfusionMap)
     assert pm1_sub.plotting.config == pm1.plotting.config
@@ -561,66 +565,66 @@ def test_mean():
 
 def test_replace_defaults():
     """Test that replacing PixelMapPlotParameters defaults works as expected."""
-    pm0 = PixelMap([[0]])
+    pm0 = PixelMap([[1]])
     assert pm0.plotting.config.cmap == "viridis"
 
     _PLOT_CONFIG_REGISTRY[PixelMap] = PixelMapPlotConfig(cmap="plasma")
 
-    pm1 = PixelMap([[0]])
+    pm1 = PixelMap([[1]])
     assert pm0.plotting.config.cmap == "viridis"
     assert pm1.plotting.config.cmap == "plasma"
 
-    pm2 = TIVMap([[0]])
+    pm2 = TIVMap([[1]])
     assert isinstance(pm2.plotting.config.cmap, Colormap)
     assert pm2.plotting.config.cmap.name == "Blues_r"
 
     _PLOT_CONFIG_REGISTRY[TIVMap] = _PLOT_CONFIG_REGISTRY[TIVMap].update(cmap="Greens")
 
-    pm3 = TIVMap([[0]])
+    pm3 = TIVMap([[1]])
     assert pm3.plotting.config.cmap == "Greens"
     assert pm3.plotting.config.colorbar
 
     _PLOT_CONFIG_REGISTRY[TIVMap] = _PLOT_CONFIG_REGISTRY[TIVMap].update(colorbar=False)
 
-    pm4 = TIVMap([[0]])
+    pm4 = TIVMap([[1]])
     assert pm4.plotting.config.cmap == "Greens"
     assert not pm4.plotting.config.colorbar
 
     reset_plot_config()
-    pm0 = PixelMap([[0]])
+    pm0 = PixelMap([[1]])
     assert pm0.plotting.config.cmap == "viridis"
 
 
 def test_set_pixelmap_plot_parameters():
     """Test that set_pixelmap_plot_parameters works as expected."""
-    pm0 = PixelMap([[0]])
[…]
