### 2025-08-25 fb303a4
Update to ruff 0.12.10
No code changes were needed.
