### 2025-08-25 2a0f156
Fix circular import issue
