### 2025-08-25 b4f3767
Deployed 32bdb8cb to v1.8.0 with MkDocs 1.6.1 and mike 2.1.3
