### 2025-08-27 EIT-ALIVE/eitprocessing#429 @ixjlyons
Link to developer doc from website broken
Minor issue in the docs. There's a link to the dev README here that appears to 404: https://eit-alive.github.io/eitprocessing/latest/contributing_doc/#you-want-to-make-some-kind-of-change-to-the-code-base

xref JOSS review: openjournals/joss-reviews#8179
