### 2025-08-27 EIT-ALIVE/eitprocessing#430 @ixjlyons
Documentation on retrieving test data
I'm not finding documentation for running the tests locally. I see from the CI config there's a Docker container and from #302 there's a dataset on Zenodo. I think a mention of this with a link in the dev README would suffice.

Since it's fairly related, I'll also suggest making it easier/possible to run the notebooks locally. Since they are rendered in the online documentation and (real) users are likely to be interested in using their own datasets, this is probably lower priority than the unit test data.

xref JOSS review: openjournals/joss-reviews#8179
