### 2025-08-27 EIT-ALIVE/eitprocessing#431 @ixjlyons
Enum members aren't listed in API docs
I think it'd be useful for the docs to list the members of enums, specifically [`Vendor`](https://eit-alive.github.io/eitprocessing/latest/api/datacontainers/eitdata/#eitprocessing.datahandling.eitdata.Vendor). `load_eit_data` docs mention vendor-specific loading functions, but those don't seem to be present in the API docs so some concrete enumeration of the built-in readers would probably be helpful.

xref JOSS review: openjournals/joss-reviews#8179
