### 2025-08-27 #8179 @ixjlyons
My apologies for the long delay in getting back to the review.

Overall the paper/docs are well written and clear, and the software itself looks good. The file readers alone are probably quite a useful resource!

I've filed a couple issues and would say the only blocker for me is EIT-ALIVE/eitprocessing#430. I don't think I have access to the test data referred to in the unit tests or the example notebooks, so I can only partially verify functionality.

Aside from this, I'll just make a couple minor comments on the paper:

- Line 49: "Standardized implementation of EIT information..." seems like there might be a missing word or something. Maybe change information to processing?
- In searching around a bit to get a sense of the state of the field and available packages for processing EIT data, I came across [pyEIT](https://github.com/eitcom/pyEIT). It appears to me to have a completely different focus (fair to say spatial rather than temporal?), but it might be worth citing and contrasting with the processing methods eitprocessing focuses on.
