+++ v1.8/examples/notebooks/ignore/test_overload.py
@@ -1,9 +1,0 @@
-from typing import Generic, TypeVar
-
-T = TypeVar("T", int, str)
-
-
-class A(Generic[T]):
-    """test."""
-
-    test: T = 3
