+++ v1.8/examples/notebooks/ignore/test_draeger_loading.py
@@ -1,5 +1,0 @@
-# %%
-
-from eitprocessing.datahandling.loading import load_eit_data
-
-seq = load_eit_data("tests/test_data/Draeger_Test.bin", vendor="draeger")
