+++ v1.8/examples/notebooks/stable_period_selection/config.yaml
@@ -1,3 +1,0 @@
-stable_period_selection:
-  selector_initials: "PS"
-  data_path: "/Volumes/Beademing/COVID-19/Studies/Stable period selection/Data"
