### 2025-08-27 addfa75
Remove build documentation on push to develop
Pushing to develop should only happen after a PR is merged. There is already a build action on PRs.
