+++ .github/workflows/test_build_documentation.yml
@@ -4,7 +4,6 @@ on:
   push:
     branches:
       - main
-      - develop
   pull_request:
     types: [opened, synchronize, reopened, ready_for_review]
     branches:
