### 2025-08-27 395b82f
Remove superfluous permissions and strategy fields
