+++ .github/workflows/test_build_documentation.yml
@@ -15,10 +15,6 @@ jobs:
     if: github.event.pull_request.draft == false
     name: Build documentation
     runs-on: ubuntu-latest
-    permissions:
-      pages: write
-    strategy:
-      fail-fast: false
     steps:
       - uses: actions/checkout@v4
         with:
