### 2025-08-27 06199d1
Add GH_PAT to checkout to properly set up permissions
