+++ .github/workflows/test_build_documentation.yml
@@ -19,6 +19,7 @@ jobs:
       - uses: actions/checkout@v4
         with:
           fetch-depth: 0
+          token: ${{ secrets.GH_PAT }}
       - uses: ./.github/actions/install_eitprocessing
         with:
           dependencies: docs
