### 2025-08-27 1dd49f6
Split off linking notebooks
