### 2025-08-27 ede5cfb
Don't build docs before build and deploy on push to main
