### 2025-08-27 4a78ad3
Rename step to reflect build and deploy
