+++ .github/workflows/test_build_documentation.yml
@@ -29,7 +29,7 @@ jobs:
       - name: Build documentation
         if: ${{ !(github.event_name == 'push' && github.ref == 'refs/heads/main') }}
         run: mkdocs build
-      - name: Deploy documentation
+      - name: Build and deploy documentation
         if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
         run: |
           git config --local user.email "github-actions[bot]@users.noreply.github.com"
