### 2025-09-05 658a2ff
Fix bug in pixel breath for NaN pixels
