### 2025-09-05 cd79a3f
Remove duplicate definition of pi variable
