### 2025-09-05 95f59a7
Set a pixel to NaN to test handling of NaNs in real data
