+++ .github/workflows/release_pypi.yml
@@ -1,54 +1,0 @@
-name: Build and upload to PyPI
-
-on:
-  workflow_dispatch:
-  release:
-    types:
-      - published
-
-jobs:
-  build:
-    name: Build universal wheel and source distribution
-    runs-on: "ubuntu-latest"
-
-    steps:
-      - uses: actions/checkout@v4
-      - uses: ./.github/actions/install_eitprocessing
-        with:
-          dependencies: publishing
-          extract-data: false
-          python-version: "3.10"
-      - name: Build wheel and source distribution
-        run: python -m build
-      - uses: actions/upload-artifact@v4
-        with:
-          path: dist/*
-
-  upload_test_pypi:
-    needs: [build]
-    runs-on: ubuntu-latest
-    if: github.event_name == 'workflow_dispatch'
-    steps:
-      - uses: actions/download-artifact@v4
-        with:
-          name: artifact
-          path: dist
-      - uses: pypa/gh-action-pypi-publish@v1.4.2
-        with:
-          user: __token__
-          password: ${{ secrets.TEST_PYPI_TOKEN_ALIVE }}
-          repository_url: https://test.pypi.org/legacy/
-
-  upload_pypi:
-    needs: [build]
-    runs-on: ubuntu-latest
-    if: github.event_name == 'release' && github.event.action == 'published'
-    steps:
-      - uses: actions/download-artifact@v4
-        with:
-          name: artifact
-          path: dist
-      - uses: pypa/gh-action-pypi-publish@v1.4.2
-        with:
-          user: __token__
-          password: ${{ secrets.PYPI_TOKEN_ALIVE }}
