+++ .github/workflows/release_github.yml
@@ -223,21 +223,3 @@ jobs:
               --title="Release v${{ needs.merge_and_bump.outputs.new-version }}" \
               --generate-notes \
               --draft
-
-  remove_branch:
-    name: Remove PR branch
-    needs: github_release
-    if: ${{ github.ref_name != 'develop' }}
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-        with:
-          ref: develop
-          fetch-depth: 0
-
-      - name: Remove PR branch
-        uses: dawidd6/action-delete-branch@v3
-        with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          branches: ${{ github.ref_name }}
