### 2025-09-16 445d4cd
Add admin flag to merge PR
