### 2025-09-22 bb74a08
Extend the documentation of the Vendor enum
Also, replace LowercaseStrEnum with Enum and remove the auto() calls for better readibility of the documentation.
