### 2025-09-22 EIT-ALIVE/eitprocessing#429 @psomhorst
This is fixed by #450.
