### 2025-09-22 933f2df
Add reference to existing EIT software packages
