+++ paper/paper.bib
@@ -285,3 +285,22 @@
   language = {en},
   doi      = {10.1186/s40635-024-00686-9}
 }
+
+
+@article{liu2018pyeit,
+  title={pyEIT: A python based framework for Electrical Impedance Tomography},
+  author={Liu, Benyuan and Yang, Bin and Xu, Canhua and Xia, Junying and Dai, Meng and Ji, Zhenyu and You, Fusheng and Dong, Xiuzhen and Shi, Xuetao and Fu, Feng},
+  journal={SoftwareX},
+  volume={7},
+  pages={304--308},
+  year={2018},
+  publisher={Elsevier}
+}
+
+@article{EIDORS2005,
+author = {Adler, Andy and Lionheart, William},
+year = {2005},
+month = {01},
+pages = {},
+title = {EIDORS: Towards a community-based extensible software base for EIT}
+}
