### 2025-09-22 7821bc3
Clarify sentences about EIT-based information use in a clinical setting
