### 2025-09-22 26db167
Add DOI for pyEIT paper
