+++ paper/paper.bib
@@ -294,7 +294,8 @@
   volume={7},
   pages={304--308},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.softx.2018.09.005}
 }
 
 @article{EIDORS2005,
