### 2025-12-18 #8179 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
