### 2025-12-29 #8179 @psomhorst
> Ready to recommend acceptance! 🎉  🥳
>
> [@ajbaird](https://github.com/ajbaird) and [@ixjlyons](https://github.com/ixjlyons): thank you again very much for your reviewing work! I and JOSS appreciate the time and effort you have put into this. 💙
>
> [@psomhorst](https://github.com/psomhorst) I will now hand this off to the EiC. In the meanwhile, I invite you to have a last look at the paper to make sure it looks as you want it to look 🙂
>
> Thanks everyone! 🌻 ⛄
>
> PS: [@psomhorst](https://github.com/psomhorst) if you haven't already, consider signing up as a reviewer 🙂 https://reviewers.joss.theoj.org/join

@britta-wstnr Thank you for your contributions and recommendation! I will look into signing up as a reviewer.
